docs(*): rewrite all docs to TSDoc and parse it with TypeDoc
