+++ docs/assets/js/search.js
@@ -1,3 +1,3 @@
 var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"2":"Module","4":"Enumeration","16":"Enumeration member","32":"Variable","64":"Function","128":"Class","256":"Interface","512":"Constructor","1024":"Property","2048":"Method","65536":"Type literal","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":256,"name":"MeasurementSystemData","url":"interfaces/_favware_converter.measurementsystemdata.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":1,"kind":1024,"name":"default","url":"interfaces/_favware_converter.measurementsystemdata.html#default","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.MeasurementSystemData"},{"id":2,"kind":1024,"name":"ratio","url":"interfaces/_favware_converter.measurementsystemdata.html#ratio","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.MeasurementSystemData"},{"id":3,"kind":2048,"name":"transform","url":"interfaces/_favware_converter.measurementsystemdata.html#transform","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"@favware/converter.MeasurementSystemData"},{"id":4,"kind":256,"name":"UnitDefinition","url":"interfaces/_favware_converter.unitdefinition.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":5,"kind":1024,"name":"name","url":"interfaces/_favware_converter.unitdefinition.html#name","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":6,"kind":1024,"name":"primary","url":"interfaces/_favware_converter.unitdefinition.html#primary","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":7,"kind":1024,"name":"secondary","url":"interfaces/_favware_converter.unitdefinition.html#secondary","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":8,"kind":1024,"name":"data","url":"interfaces/_favware_converter.unitdefinition.html#data","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":9,"kind":256,"name":"UnitData","url":"interfaces/_favware_converter.unitdata.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":10,"kind":1024,"name":"id","url":"interfaces/_favware_converter.unitdata.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":11,"kind":1024,"name":"system","url":"interfaces/_favware_converter.unitdata.html#system","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":12,"kind":1024,"name":"multiplier","url":"interfaces/_favware_converter.unitdata.html#multiplier","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":13,"kind":1024,"name":"valueShift","url":"interfaces/_favware_converter.unitdata.html#valueshift","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":14,"kind":256,"name":"ConvertOptions","url":"interfaces/_favware_converter.convertoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":15,"kind":1024,"name":"precision","url":"interfaces/_favware_converter.convertoptions.html#precision","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.ConvertOptions"},{"id":16,"kind":4,"name":"System","url":"enums/_favware_converter.system.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/converter"},{"id":17,"kind":16,"name":"PRIMARY","url":"enums/_favware_converter.system.html#primary","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/converter.System"},{"id":18,"kind":16,"name":"SECONDARY","url":"enums/_favware_converter.system.html#secondary","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/converter.System"},{"id":19,"kind":128,"name":"ConverterError","url":"classes/_favware_converter.convertererror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/converter"},{"id":20,"kind":512,"name":"constructor","url":"classes/_favware_converter.convertererror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"@favware/converter.ConverterError"},{"id":21,"kind":1024,"name":"name","url":"classes/_favware_converter.convertererror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/converter.ConverterError"},{"id":22,"kind":1024,"name":"message","url":"classes/_favware_converter.convertererror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/converter.ConverterError"},{"id":23,"kind":1024,"name":"stack","url":"classes/_favware_converter.convertererror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/converter.ConverterError"},{"id":24,"kind":1024,"name":"Error","url":"classes/_favware_converter.convertererror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/converter.ConverterError"},{"id":25,"kind":32,"name":"definitions","url":"modules/_favware_converter.html#definitions","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/converter"},{"id":26,"kind":64,"name":"roundNumber","url":"modules/_favware_converter.html#roundnumber","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/converter"},{"id":27,"kind":64,"name":"convert","url":"modules/_favware_converter.html#convert","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/converter"},{"id":28,"kind":4,"name":"Radix","url":"enums/_favware_crypto.radix.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/crypto"},{"id":29,"kind":16,"name":"base16","url":"enums/_favware_crypto.radix.html#base16","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":30,"kind":16,"name":"base32","url":"enums/_favware_crypto.radix.html#base32","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":31,"kind":16,"name":"base36","url":"enums/_favware_crypto.radix.html#base36","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":32,"kind":16,"name":"base58","url":"enums/_favware_crypto.radix.html#base58","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":33,"kind":16,"name":"base64","url":"enums/_favware_crypto.radix.html#base64","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":34,"kind":128,"name":"CryptoError","url":"classes/_favware_crypto.cryptoerror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/crypto"},{"id":35,"kind":512,"name":"constructor","url":"classes/_favware_crypto.cryptoerror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"@favware/crypto.CryptoError"},{"id":36,"kind":1024,"name":"name","url":"classes/_favware_crypto.cryptoerror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/crypto.CryptoError"},{"id":37,"kind":1024,"name":"message","url":"classes/_favware_crypto.cryptoerror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/crypto.CryptoError"},{"id":38,"kind":1024,"name":"stack","url":"classes/_favware_crypto.cryptoerror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/crypto.CryptoError"},{"id":39,"kind":1024,"name":"Error","url":"classes/_favware_crypto.cryptoerror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/crypto.CryptoError"},{"id":40,"kind":32,"name":"DEFAULT_BITS","url":"modules/_favware_crypto.html#default_bits","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/crypto"},{"id":41,"kind":32,"name":"DEFAULT_RADIX","url":"modules/_favware_crypto.html#default_radix","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/crypto"},{"id":42,"kind":32,"name":"CHARACTERS","url":"modules/_favware_crypto.html#characters","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/crypto"},{"id":43,"kind":64,"name":"crypto","url":"modules/_favware_crypto.html#crypto","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/crypto"},{"id":44,"kind":2097152,"name":"config","url":"modules/_favware_eslint_config.html#config","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/eslint-config"},{"id":45,"kind":32,"name":"root","url":"modules/_favware_eslint_config.html#config.root","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":46,"kind":32,"name":"parser","url":"modules/_favware_eslint_config.html#config.parser","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":47,"kind":2097152,"name":"parserOptions","url":"modules/_favware_eslint_config.html#config.parseroptions","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":48,"kind":32,"name":"project","url":"modules/_favware_eslint_config.html#config.parseroptions.project","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":49,"kind":32,"name":"jsx","url":"modules/_favware_eslint_config.html#config.parseroptions.jsx","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":50,"kind":32,"name":"sourceType","url":"modules/_favware_eslint_config.html#config.parseroptions.sourcetype","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":51,"kind":32,"name":"ecmaVersion","url":"modules/_favware_eslint_config.html#config.parseroptions.ecmaversion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":52,"kind":2097152,"name":"ecmaFeatures","url":"modules/_favware_eslint_config.html#config.parseroptions.ecmafeatures","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":53,"kind":32,"name":"jsx","url":"modules/_favware_eslint_config.html#config.parseroptions.ecmafeatures.jsx-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions.ecmaFeatures"},{"id":54,"kind":32,"name":"extends","url":"modules/_favware_eslint_config.html#config.extends","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":55,"kind":2097152,"name":"env","url":"modules/_favware_eslint_config.html#config.env","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":56,"kind":32,"name":"node","url":"modules/_favware_eslint_config.html#config.env.node","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":57,"kind":32,"name":"es6","url":"modules/_favware_eslint_config.html#config.env.es6","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":58,"kind":32,"name":"jest","url":"modules/_favware_eslint_config.html#config.env.jest","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":59,"kind":32,"name":"browser","url":"modules/_favware_eslint_config.html#config.env.browser","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":60,"kind":32,"name":"commonjs","url":"modules/_favware_eslint_config.html#config.env.commonjs","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":61,"kind":2097152,"name":"rules","url":"modules/_favware_eslint_config.html#config.rules","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":62,"kind":32,"name":"@typescript-eslint/indent","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_indent","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":63,"kind":32,"name":"@typescript-eslint/no-non-null-assertion","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_non_null_assertion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":64,"kind":32,"name":"@typescript-eslint/explicit-function-return-type","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_explicit_function_return_type","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":65,"kind":32,"name":"@typescript-eslint/prefer-interface","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_prefer_interface","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":66,"kind":32,"name":"@typescript-eslint/consistent-type-definitions","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_consistent_type_definitions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":67,"kind":32,"name":"@typescript-eslint/interface-name-prefix","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_interface_name_prefix","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":68,"kind":32,"name":"@typescript-eslint/await-thenable","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_await_thenable","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":69,"kind":32,"name":"@typescript-eslint/ban-types","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_ban_types","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":70,"kind":32,"name":"@typescript-eslint/func-call-spacing","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_func_call_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":71,"kind":32,"name":"@typescript-eslint/member-ordering","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_member_ordering","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":72,"kind":32,"name":"@typescript-eslint/no-empty-function","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_empty_function","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":73,"kind":32,"name":"@typescript-eslint/no-extra-parens","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_extra_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":74,"kind":32,"name":"@typescript-eslint/no-extraneous-class","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_extraneous_class","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":75,"kind":32,"name":"@typescript-eslint/no-for-in-array","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_for_in_array","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":76,"kind":32,"name":"@typescript-eslint/no-require-imports","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_require_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":77,"kind":32,"name":"@typescript-eslint/no-this-alias","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_this_alias","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":78,"kind":32,"name":"@typescript-eslint/no-unnecessary-qualifier","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_unnecessary_qualifier","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":79,"kind":32,"name":"@typescript-eslint/no-unnecessary-type-assertion","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_unnecessary_type_assertion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":80,"kind":32,"name":"@typescript-eslint/no-useless-constructor","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_useless_constructor","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":81,"kind":32,"name":"@typescript-eslint/prefer-for-of","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_prefer_for_of","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":82,"kind":32,"name":"@typescript-eslint/prefer-includes","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_prefer_includes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":83,"kind":32,"name":"@typescript-eslint/promise-function-async","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_promise_function_async","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":84,"kind":32,"name":"@typescript-eslint/require-array-sort-compare","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_require_array_sort_compare","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":85,"kind":32,"name":"@typescript-eslint/restrict-plus-operands","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_restrict_plus_operands","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":86,"kind":32,"name":"@typescript-eslint/semi","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_semi","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":87,"kind":32,"name":"@typescript-eslint/unbound-method","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_unbound_method","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":88,"kind":32,"name":"@typescript-eslint/unified-signatures","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_unified_signatures","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":89,"kind":32,"name":"@typescript-eslint/camelcase","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_camelcase","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":90,"kind":32,"name":"jest/consistent-test-it","url":"modules/_favware_eslint_config.html#config.rules.jest_consistent_test_it","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":91,"kind":32,"name":"jest/no-hooks","url":"modules/_favware_eslint_config.html#config.rules.jest_no_hooks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":92,"kind":32,"name":"jest/lowercase-name","url":"modules/_favware_eslint_config.html#config.rules.jest_lowercase_name","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":93,"kind":32,"name":"jest/no-export","url":"modules/_favware_eslint_config.html#config.rules.jest_no_export","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":94,"kind":32,"name":"jest/prefer-expect-assertions","url":"modules/_favware_eslint_config.html#config.rules.jest_prefer_expect_assertions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":95,"kind":32,"name":"jest/no-if","url":"modules/_favware_eslint_config.html#config.rules.jest_no_if","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":96,"kind":32,"name":"jest/prefer-inline-snapshots","url":"modules/_favware_eslint_config.html#config.rules.jest_prefer_inline_snapshots","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":97,"kind":32,"name":"jest/expect-expect","url":"modules/_favware_eslint_config.html#config.rules.jest_expect_expect","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":98,"kind":32,"name":"jest/require-top-level-describe","url":"modules/_favware_eslint_config.html#config.rules.jest_require_top_level_describe","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":99,"kind":32,"name":"no-console","url":"modules/_favware_eslint_config.html#config.rules.no_console","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":100,"kind":32,"name":"no-template-curly-in-string","url":"modules/_favware_eslint_config.html#config.rules.no_template_curly_in_string","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":101,"kind":32,"name":"require-atomic-updates","url":"modules/_favware_eslint_config.html#config.rules.require_atomic_updates","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":102,"kind":32,"name":"accessor-pairs","url":"modules/_favware_eslint_config.html#config.rules.accessor_pairs","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":103,"kind":32,"name":"array-callback-return","url":"modules/_favware_eslint_config.html#config.rules.array_callback_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":104,"kind":32,"name":"block-scoped-var","url":"modules/_favware_eslint_config.html#config.rules.block_scoped_var","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":105,"kind":32,"name":"consistent-return","url":"modules/_favware_eslint_config.html#config.rules.consistent_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":106,"kind":32,"name":"curly","url":"modules/_favware_eslint_config.html#config.rules.curly","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":107,"kind":32,"name":"default-case","url":"modules/_favware_eslint_config.html#config.rules.default_case","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":108,"kind":32,"name":"dot-location","url":"modules/_favware_eslint_config.html#config.rules.dot_location","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":109,"kind":32,"name":"eqeqeq","url":"modules/_favware_eslint_config.html#config.rules.eqeqeq","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":110,"kind":32,"name":"max-classes-per-file","url":"modules/_favware_eslint_config.html#config.rules.max_classes_per_file","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":111,"kind":32,"name":"no-alert","url":"modules/_favware_eslint_config.html#config.rules.no_alert","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":112,"kind":32,"name":"no-caller","url":"modules/_favware_eslint_config.html#config.rules.no_caller","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":113,"kind":32,"name":"no-div-regex","url":"modules/_favware_eslint_config.html#config.rules.no_div_regex","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":114,"kind":32,"name":"no-else-return","url":"modules/_favware_eslint_config.html#config.rules.no_else_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":115,"kind":32,"name":"no-eq-null","url":"modules/_favware_eslint_config.html#config.rules.no_eq_null","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":116,"kind":32,"name":"no-eval","url":"modules/_favware_eslint_config.html#config.rules.no_eval","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":117,"kind":32,"name":"no-extend-native","url":"modules/_favware_eslint_config.html#config.rules.no_extend_native","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":118,"kind":32,"name":"no-extra-bind","url":"modules/_favware_eslint_config.html#config.rules.no_extra_bind","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":119,"kind":32,"name":"no-extra-label","url":"modules/_favware_eslint_config.html#config.rules.no_extra_label","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":120,"kind":32,"name":"no-floating-decimal","url":"modules/_favware_eslint_config.html#config.rules.no_floating_decimal","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":121,"kind":32,"name":"no-implicit-coercion","url":"modules/_favware_eslint_config.html#config.rules.no_implicit_coercion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":122,"kind":32,"name":"no-implicit-globals","url":"modules/_favware_eslint_config.html#config.rules.no_implicit_globals","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":123,"kind":32,"name":"no-implied-eval","url":"modules/_favware_eslint_config.html#config.rules.no_implied_eval","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":124,"kind":32,"name":"no-invalid-this","url":"modules/_favware_eslint_config.html#config.rules.no_invalid_this","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":125,"kind":32,"name":"no-iterator","url":"modules/_favware_eslint_config.html#config.rules.no_iterator","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":126,"kind":32,"name":"no-lone-blocks","url":"modules/_favware_eslint_config.html#config.rules.no_lone_blocks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":127,"kind":32,"name":"no-loop-func","url":"modules/_favware_eslint_config.html#config.rules.no_loop_func","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":128,"kind":32,"name":"no-multi-spaces","url":"modules/_favware_eslint_config.html#config.rules.no_multi_spaces","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":129,"kind":32,"name":"no-multi-str","url":"modules/_favware_eslint_config.html#config.rules.no_multi_str","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":130,"kind":32,"name":"no-new","url":"modules/_favware_eslint_config.html#config.rules.no_new","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":131,"kind":32,"name":"no-new-func","url":"modules/_favware_eslint_config.html#config.rules.no_new_func","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":132,"kind":32,"name":"no-new-wrappers","url":"modules/_favware_eslint_config.html#config.rules.no_new_wrappers","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":133,"kind":32,"name":"no-octal-escape","url":"modules/_favware_eslint_config.html#config.rules.no_octal_escape","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":134,"kind":32,"name":"no-proto","url":"modules/_favware_eslint_config.html#config.rules.no_proto","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":135,"kind":32,"name":"no-prototype-builtins","url":"modules/_favware_eslint_config.html#config.rules.no_prototype_builtins","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":136,"kind":32,"name":"no-return-assign","url":"modules/_favware_eslint_config.html#config.rules.no_return_assign","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":137,"kind":32,"name":"no-return-await","url":"modules/_favware_eslint_config.html#config.rules.no_return_await","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":138,"kind":32,"name":"no-script-url","url":"modules/_favware_eslint_config.html#config.rules.no_script_url","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":139,"kind":32,"name":"no-self-compare","url":"modules/_favware_eslint_config.html#config.rules.no_self_compare","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":140,"kind":32,"name":"no-sequences","url":"modules/_favware_eslint_config.html#config.rules.no_sequences","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":141,"kind":32,"name":"no-throw-literal","url":"modules/_favware_eslint_config.html#config.rules.no_throw_literal","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":142,"kind":32,"name":"no-unmodified-loop-condition","url":"modules/_favware_eslint_config.html#config.rules.no_unmodified_loop_condition","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":143,"kind":32,"name":"no-unused-expressions","url":"modules/_favware_eslint_config.html#config.rules.no_unused_expressions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":144,"kind":32,"name":"no-useless-call","url":"modules/_favware_eslint_config.html#config.rules.no_useless_call","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":145,"kind":32,"name":"no-useless-concat","url":"modules/_favware_eslint_config.html#config.rules.no_useless_concat","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":146,"kind":32,"name":"no-useless-return","url":"modules/_favware_eslint_config.html#config.rules.no_useless_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":147,"kind":32,"name":"no-void","url":"modules/_favware_eslint_config.html#config.rules.no_void","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":148,"kind":32,"name":"no-warning-comments","url":"modules/_favware_eslint_config.html#config.rules.no_warning_comments","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":149,"kind":32,"name":"prefer-promise-reject-errors","url":"modules/_favware_eslint_config.html#config.rules.prefer_promise_reject_errors","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":150,"kind":32,"name":"radix","url":"modules/_favware_eslint_config.html#config.rules.radix","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":151,"kind":32,"name":"wrap-iife","url":"modules/_favware_eslint_config.html#config.rules.wrap_iife","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":152,"kind":32,"name":"yoda","url":"modules/_favware_eslint_config.html#config.rules.yoda","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":153,"kind":32,"name":"strict","url":"modules/_favware_eslint_config.html#config.rules.strict","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":154,"kind":32,"name":"no-label-var","url":"modules/_favware_eslint_config.html#config.rules.no_label_var","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":155,"kind":32,"name":"no-shadow","url":"modules/_favware_eslint_config.html#config.rules.no_shadow","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":156,"kind":32,"name":"no-undef-init","url":"modules/_favware_eslint_config.html#config.rules.no_undef_init","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":157,"kind":32,"name":"callback-return","url":"modules/_favware_eslint_config.html#config.rules.callback_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":158,"kind":32,"name":"global-require","url":"modules/_favware_eslint_config.html#config.rules.global_require","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":159,"kind":32,"name":"handle-callback-err","url":"modules/_favware_eslint_config.html#config.rules.handle_callback_err","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":160,"kind":32,"name":"no-buffer-constructor","url":"modules/_favware_eslint_config.html#config.rules.no_buffer_constructor","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":161,"kind":32,"name":"no-mixed-requires","url":"modules/_favware_eslint_config.html#config.rules.no_mixed_requires","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":162,"kind":32,"name":"no-new-require","url":"modules/_favware_eslint_config.html#config.rules.no_new_require","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":163,"kind":32,"name":"no-path-concat","url":"modules/_favware_eslint_config.html#config.rules.no_path_concat","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":164,"kind":32,"name":"array-bracket-spacing","url":"modules/_favware_eslint_config.html#config.rules.array_bracket_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":165,"kind":32,"name":"block-spacing","url":"modules/_favware_eslint_config.html#config.rules.block_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":166,"kind":32,"name":"brace-style","url":"modules/_favware_eslint_config.html#config.rules.brace_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":167,"kind":32,"name":"capitalized-comments","url":"modules/_favware_eslint_config.html#config.rules.capitalized_comments","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":168,"kind":32,"name":"comma-dangle","url":"modules/_favware_eslint_config.html#config.rules.comma_dangle","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":169,"kind":32,"name":"comma-spacing","url":"modules/_favware_eslint_config.html#config.rules.comma_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":170,"kind":32,"name":"comma-style","url":"modules/_favware_eslint_config.html#config.rules.comma_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":171,"kind":32,"name":"computed-property-spacing","url":"modules/_favware_eslint_config.html#config.rules.computed_property_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":172,"kind":32,"name":"eol-last","url":"modules/_favware_eslint_config.html#config.rules.eol_last","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":173,"kind":32,"name":"func-name-matching","url":"modules/_favware_eslint_config.html#config.rules.func_name_matching","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":174,"kind":32,"name":"func-names","url":"modules/_favware_eslint_config.html#config.rules.func_names","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":175,"kind":32,"name":"func-style","url":"modules/_favware_eslint_config.html#config.rules.func_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":176,"kind":32,"name":"implicit-arrow-linebreak","url":"modules/_favware_eslint_config.html#config.rules.implicit_arrow_linebreak","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":177,"kind":32,"name":"jsx-quotes","url":"modules/_favware_eslint_config.html#config.rules.jsx_quotes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":178,"kind":32,"name":"key-spacing","url":"modules/_favware_eslint_config.html#config.rules.key_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":179,"kind":32,"name":"keyword-spacing","url":"modules/_favware_eslint_config.html#config.rules.keyword_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":180,"kind":32,"name":"linebreak-style","url":"modules/_favware_eslint_config.html#config.rules.linebreak_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":181,"kind":32,"name":"lines-between-class-members","url":"modules/_favware_eslint_config.html#config.rules.lines_between_class_members","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":182,"kind":32,"name":"max-len","url":"modules/_favware_eslint_config.html#config.rules.max_len","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":183,"kind":32,"name":"max-statements-per-line","url":"modules/_favware_eslint_config.html#config.rules.max_statements_per_line","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":184,"kind":32,"name":"new-cap","url":"modules/_favware_eslint_config.html#config.rules.new_cap","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":185,"kind":32,"name":"new-parens","url":"modules/_favware_eslint_config.html#config.rules.new_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":186,"kind":32,"name":"no-array-constructor","url":"modules/_favware_eslint_config.html#config.rules.no_array_constructor","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":187,"kind":32,"name":"no-lonely-if","url":"modules/_favware_eslint_config.html#config.rules.no_lonely_if","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":188,"kind":32,"name":"no-mixed-operators","url":"modules/_favware_eslint_config.html#config.rules.no_mixed_operators","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":189,"kind":32,"name":"no-multi-assign","url":"modules/_favware_eslint_config.html#config.rules.no_multi_assign","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":190,"kind":32,"name":"no-multiple-empty-lines","url":"modules/_favware_eslint_config.html#config.rules.no_multiple_empty_lines","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":191,"kind":32,"name":"no-nested-ternary","url":"modules/_favware_eslint_config.html#config.rules.no_nested_ternary","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":192,"kind":32,"name":"no-new-object","url":"modules/_favware_eslint_config.html#config.rules.no_new_object","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":193,"kind":32,"name":"no-tabs","url":"modules/_favware_eslint_config.html#config.rules.no_tabs","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":194,"kind":32,"name":"no-trailing-spaces","url":"modules/_favware_eslint_config.html#config.rules.no_trailing_spaces","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":195,"kind":32,"name":"no-underscore-dangle","url":"modules/_favware_eslint_config.html#config.rules.no_underscore_dangle","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":196,"kind":32,"name":"no-unneeded-ternary","url":"modules/_favware_eslint_config.html#config.rules.no_unneeded_ternary","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":197,"kind":32,"name":"no-whitespace-before-property","url":"modules/_favware_eslint_config.html#config.rules.no_whitespace_before_property","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":198,"kind":32,"name":"nonblock-statement-body-position","url":"modules/_favware_eslint_config.html#config.rules.nonblock_statement_body_position","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":199,"kind":32,"name":"object-curly-newline","url":"modules/_favware_eslint_config.html#config.rules.object_curly_newline","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":200,"kind":32,"name":"one-var","url":"modules/_favware_eslint_config.html#config.rules.one_var","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":201,"kind":32,"name":"one-var-declaration-per-line","url":"modules/_favware_eslint_config.html#config.rules.one_var_declaration_per_line","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":202,"kind":32,"name":"operator-linebreak","url":"modules/_favware_eslint_config.html#config.rules.operator_linebreak","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":203,"kind":32,"name":"padded-blocks","url":"modules/_favware_eslint_config.html#config.rules.padded_blocks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":204,"kind":32,"name":"padding-line-between-statements","url":"modules/_favware_eslint_config.html#config.rules.padding_line_between_statements","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":205,"kind":32,"name":"prefer-object-spread","url":"modules/_favware_eslint_config.html#config.rules.prefer_object_spread","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":206,"kind":32,"name":"quote-props","url":"modules/_favware_eslint_config.html#config.rules.quote_props","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":207,"kind":32,"name":"quotes","url":"modules/_favware_eslint_config.html#config.rules.quotes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":208,"kind":32,"name":"semi-spacing","url":"modules/_favware_eslint_config.html#config.rules.semi_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":209,"kind":32,"name":"semi-style","url":"modules/_favware_eslint_config.html#config.rules.semi_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":210,"kind":32,"name":"sort-vars","url":"modules/_favware_eslint_config.html#config.rules.sort_vars","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":211,"kind":32,"name":"space-before-blocks","url":"modules/_favware_eslint_config.html#config.rules.space_before_blocks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":212,"kind":32,"name":"space-before-function-paren","url":"modules/_favware_eslint_config.html#config.rules.space_before_function_paren","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":213,"kind":32,"name":"space-in-parens","url":"modules/_favware_eslint_config.html#config.rules.space_in_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":214,"kind":32,"name":"space-infix-ops","url":"modules/_favware_eslint_config.html#config.rules.space_infix_ops","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":215,"kind":32,"name":"space-unary-ops","url":"modules/_favware_eslint_config.html#config.rules.space_unary_ops","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":216,"kind":32,"name":"switch-colon-spacing","url":"modules/_favware_eslint_config.html#config.rules.switch_colon_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":217,"kind":32,"name":"template-tag-spacing","url":"modules/_favware_eslint_config.html#config.rules.template_tag_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":218,"kind":32,"name":"no-useless-computed-key","url":"modules/_favware_eslint_config.html#config.rules.no_useless_computed_key","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":219,"kind":32,"name":"arrow-parens","url":"modules/_favware_eslint_config.html#config.rules.arrow_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":220,"kind":32,"name":"arrow-spacing","url":"modules/_favware_eslint_config.html#config.rules.arrow_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":221,"kind":32,"name":"generator-star-spacing","url":"modules/_favware_eslint_config.html#config.rules.generator_star_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":222,"kind":32,"name":"no-duplicate-imports","url":"modules/_favware_eslint_config.html#config.rules.no_duplicate_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":223,"kind":32,"name":"no-useless-rename","url":"modules/_favware_eslint_config.html#config.rules.no_useless_rename","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":224,"kind":32,"name":"prefer-arrow-callback","url":"modules/_favware_eslint_config.html#config.rules.prefer_arrow_callback","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":225,"kind":32,"name":"prefer-numeric-literals","url":"modules/_favware_eslint_config.html#config.rules.prefer_numeric_literals","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":226,"kind":32,"name":"prefer-template","url":"modules/_favware_eslint_config.html#config.rules.prefer_template","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":227,"kind":32,"name":"rest-spread-spacing","url":"modules/_favware_eslint_config.html#config.rules.rest_spread_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":228,"kind":32,"name":"template-curly-spacing","url":"modules/_favware_eslint_config.html#config.rules.template_curly_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":229,"kind":32,"name":"yield-star-spacing","url":"modules/_favware_eslint_config.html#config.rules.yield_star_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":230,"kind":256,"name":"PluginOptions","url":"interfaces/_favware_milky_tslint.pluginoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":231,"kind":1024,"name":"configuration","url":"interfaces/_favware_milky_tslint.pluginoptions.html#configuration","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":232,"kind":1024,"name":"fix","url":"interfaces/_favware_milky_tslint.pluginoptions.html#fix","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":233,"kind":1024,"name":"quiet","url":"interfaces/_favware_milky_tslint.pluginoptions.html#quiet","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":234,"kind":1024,"name":"formatter","url":"interfaces/_favware_milky_tslint.pluginoptions.html#formatter","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":235,"kind":1024,"name":"formattersDirectory","url":"interfaces/_favware_milky_tslint.pluginoptions.html#formattersdirectory","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":236,"kind":1024,"name":"rulesDirectory","url":"interfaces/_favware_milky_tslint.pluginoptions.html#rulesdirectory","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":237,"kind":1024,"name":"tslint","url":"interfaces/_favware_milky_tslint.pluginoptions.html#tslint","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":238,"kind":1024,"name":"program","url":"interfaces/_favware_milky_tslint.pluginoptions.html#program","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":239,"kind":256,"name":"ReportOptions","url":"interfaces/_favware_milky_tslint.reportoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":240,"kind":1024,"name":"emitError","url":"interfaces/_favware_milky_tslint.reportoptions.html#emiterror","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":241,"kind":1024,"name":"reportLimit","url":"interfaces/_favware_milky_tslint.reportoptions.html#reportlimit","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":242,"kind":1024,"name":"summarizeFailureOutput","url":"interfaces/_favware_milky_tslint.reportoptions.html#summarizefailureoutput","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":243,"kind":1024,"name":"allowWarnings","url":"interfaces/_favware_milky_tslint.reportoptions.html#allowwarnings","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":244,"kind":256,"name":"TSLintFile","url":"interfaces/_favware_milky_tslint.tslintfile.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":245,"kind":1024,"name":"tslint","url":"interfaces/_favware_milky_tslint.tslintfile.html#tslint","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":246,"kind":1024,"name":"path","url":"interfaces/_favware_milky_tslint.tslintfile.html#path","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":247,"kind":1024,"name":"relative","url":"interfaces/_favware_milky_tslint.tslintfile.html#relative","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":248,"kind":1024,"name":"contents","url":"interfaces/_favware_milky_tslint.tslintfile.html#contents","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":249,"kind":2048,"name":"isStream","url":"interfaces/_favware_milky_tslint.tslintfile.html#isstream","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":250,"kind":2048,"name":"isNull","url":"interfaces/_favware_milky_tslint.tslintfile.html#isnull","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":251,"kind":4,"name":"LogLevels","url":"enums/_favware_milky_tslint.loglevels.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":252,"kind":16,"name":"ERROR","url":"enums/_favware_milky_tslint.loglevels.html#error","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LogLevels"},{"id":253,"kind":16,"name":"INFO","url":"enums/_favware_milky_tslint.loglevels.html#info","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LogLevels"},{"id":254,"kind":16,"name":"WARN","url":"enums/_favware_milky_tslint.loglevels.html#warn","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LogLevels"},{"id":255,"kind":4,"name":"LintFormatters","url":"enums/_favware_milky_tslint.lintformatters.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":256,"kind":16,"name":"CHECKSTYLE","url":"enums/_favware_milky_tslint.lintformatters.html#checkstyle","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":257,"kind":16,"name":"CODEFRAME","url":"enums/_favware_milky_tslint.lintformatters.html#codeframe","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":258,"kind":16,"name":"FILESLIST","url":"enums/_favware_milky_tslint.lintformatters.html#fileslist","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":259,"kind":16,"name":"JSON","url":"enums/_favware_milky_tslint.lintformatters.html#json","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":260,"kind":16,"name":"JUNIT","url":"enums/_favware_milky_tslint.lintformatters.html#junit","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":261,"kind":16,"name":"MSBUILD","url":"enums/_favware_milky_tslint.lintformatters.html#msbuild","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":262,"kind":16,"name":"PMD","url":"enums/_favware_milky_tslint.lintformatters.html#pmd","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":263,"kind":16,"name":"PROSE","url":"enums/_favware_milky_tslint.lintformatters.html#prose","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":264,"kind":16,"name":"STYLISH","url":"enums/_favware_milky_tslint.lintformatters.html#stylish","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":265,"kind":16,"name":"TAP","url":"enums/_favware_milky_tslint.lintformatters.html#tap","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":266,"kind":16,"name":"VERBOSE","url":"enums/_favware_milky_tslint.lintformatters.html#verbose","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":267,"kind":16,"name":"VSO","url":"enums/_favware_milky_tslint.lintformatters.html#vso","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":268,"kind":256,"name":"MilkyEvents","url":"interfaces/_favware_milky_tslint.milkyevents.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":269,"kind":1024,"name":"data","url":"interfaces/_favware_milky_tslint.milkyevents.html#data","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.MilkyEvents"},{"id":270,"kind":65536,"name":"__type","url":"interfaces/_favware_milky_tslint.milkyevents.html#data.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"@favware/milky-tslint.MilkyEvents.data"},{"id":271,"kind":1024,"name":"error","url":"interfaces/_favware_milky_tslint.milkyevents.html#error","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.MilkyEvents"},{"id":272,"kind":65536,"name":"__type","url":"interfaces/_favware_milky_tslint.milkyevents.html#error.__type-2","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"@favware/milky-tslint.MilkyEvents.error"},{"id":273,"kind":1024,"name":"end","url":"interfaces/_favware_milky_tslint.milkyevents.html#end","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.MilkyEvents"},{"id":274,"kind":65536,"name":"__type","url":"interfaces/_favware_milky_tslint.milkyevents.html#end.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"@favware/milky-tslint.MilkyEvents.end"},{"id":275,"kind":32,"name":"eventemitter","url":"modules/_favware_milky_tslint.html#eventemitter","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":276,"kind":64,"name":"isString","url":"modules/_favware_milky_tslint.html#isstring","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":277,"kind":64,"name":"milkyLint","url":"modules/_favware_milky_tslint.html#milkylint","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":278,"kind":64,"name":"milkyReport","url":"modules/_favware_milky_tslint.html#milkyreport","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":279,"kind":256,"name":"StringifyOptions","url":"interfaces/_favware_querystring.stringifyoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/querystring"},{"id":280,"kind":1024,"name":"separator","url":"interfaces/_favware_querystring.stringifyoptions.html#separator","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.StringifyOptions"},{"id":281,"kind":1024,"name":"equals","url":"interfaces/_favware_querystring.stringifyoptions.html#equals","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.StringifyOptions"},{"id":282,"kind":1024,"name":"includeQuestion","url":"interfaces/_favware_querystring.stringifyoptions.html#includequestion","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.StringifyOptions"},{"id":283,"kind":256,"name":"ParseOptions","url":"interfaces/_favware_querystring.parseoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/querystring"},{"id":284,"kind":1024,"name":"separator","url":"interfaces/_favware_querystring.parseoptions.html#separator","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.ParseOptions"},{"id":285,"kind":1024,"name":"equals","url":"interfaces/_favware_querystring.parseoptions.html#equals","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.ParseOptions"},{"id":286,"kind":256,"name":"QuerystringObject","url":"interfaces/_favware_querystring.querystringobject.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/querystring"},{"id":287,"kind":4,"name":"FunctionSource","url":"enums/_favware_querystring.functionsource.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/querystring"},{"id":288,"kind":16,"name":"Parse","url":"enums/_favware_querystring.functionsource.html#parse","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/querystring.FunctionSource"},{"id":289,"kind":16,"name":"Stringify","url":"enums/_favware_querystring.functionsource.html#stringify","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/querystring.FunctionSource"},{"id":290,"kind":4194304,"name":"QuerystringError","url":"modules/_favware_querystring.html#querystringerror","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"@favware/querystring"},{"id":291,"kind":64,"name":"stringify","url":"modules/_favware_querystring.html#stringify","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/querystring"},{"id":292,"kind":64,"name":"parse","url":"modules/_favware_querystring.html#parse","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/querystring"},{"id":293,"kind":2097152,"name":"charSets","url":"modules/_favware_unescape.html#charsets","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/unescape"},{"id":294,"kind":32,"name":"&quot;","url":"modules/_favware_unescape.html#charsets._quot_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":295,"kind":32,"name":"&#34;","url":"modules/_favware_unescape.html#charsets.__34_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":296,"kind":32,"name":"&#034;","url":"modules/_favware_unescape.html#charsets.__034_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":297,"kind":32,"name":"&#x22;","url":"modules/_favware_unescape.html#charsets.__x22_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":298,"kind":32,"name":"&apos;","url":"modules/_favware_unescape.html#charsets._apos_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":299,"kind":32,"name":"&#39;","url":"modules/_favware_unescape.html#charsets.__39_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":300,"kind":32,"name":"&#039;","url":"modules/_favware_unescape.html#charsets.__039_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":301,"kind":32,"name":"&#x27;","url":"modules/_favware_unescape.html#charsets.__x27_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":302,"kind":32,"name":"&amp;","url":"modules/_favware_unescape.html#charsets._amp_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":303,"kind":32,"name":"&#38;","url":"modules/_favware_unescape.html#charsets.__38_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":304,"kind":32,"name":"&#038;","url":"modules/_favware_unescape.html#charsets.__038_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":305,"kind":32,"name":"&#x26;","url":"modules/_favware_unescape.html#charsets.__x26_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":306,"kind":32,"name":"&gt;","url":"modules/_favware_unescape.html#charsets._gt_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":307,"kind":32,"name":"&#62;","url":"modules/_favware_unescape.html#charsets.__62_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":308,"kind":32,"name":"&#062;","url":"modules/_favware_unescape.html#charsets.__062_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":309,"kind":32,"name":"&#x3e;","url":"modules/_favware_unescape.html#charsets.__x3e_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":310,"kind":32,"name":"&lt;","url":"modules/_favware_unescape.html#charsets._lt_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":311,"kind":32,"name":"&#60;","url":"modules/_favware_unescape.html#charsets.__60_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":312,"kind":32,"name":"&#060;","url":"modules/_favware_unescape.html#charsets.__060_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":313,"kind":32,"name":"&#x3c;","url":"modules/_favware_unescape.html#charsets.__x3c_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":314,"kind":32,"name":"&copy;","url":"modules/_favware_unescape.html#charsets._copy_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":315,"kind":32,"name":"&#169;","url":"modules/_favware_unescape.html#charsets.__169_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":316,"kind":32,"name":"&#xa9;","url":"modules/_favware_unescape.html#charsets.__xa9_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":317,"kind":32,"name":"&reg;","url":"modules/_favware_unescape.html#charsets._reg_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":318,"kind":32,"name":"&#174;","url":"modules/_favware_unescape.html#charsets.__174_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":319,"kind":32,"name":"&#xae;","url":"modules/_favware_unescape.html#charsets.__xae_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":320,"kind":32,"name":"&cent;","url":"modules/_favware_unescape.html#charsets._cent_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":321,"kind":32,"name":"&#162;","url":"modules/_favware_unescape.html#charsets.__162_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":322,"kind":32,"name":"&#xa2;","url":"modules/_favware_unescape.html#charsets.__xa2_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":323,"kind":32,"name":"&euro;","url":"modules/_favware_unescape.html#charsets._euro_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":324,"kind":32,"name":"&#8364;","url":"modules/_favware_unescape.html#charsets.__8364_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":325,"kind":32,"name":"&pound;","url":"modules/_favware_unescape.html#charsets._pound_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":326,"kind":32,"name":"&#163;","url":"modules/_favware_unescape.html#charsets.__163_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":327,"kind":32,"name":"&#xa3;","url":"modules/_favware_unescape.html#charsets.__xa3_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":328,"kind":32,"name":"&yen;","url":"modules/_favware_unescape.html#charsets._yen_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":329,"kind":32,"name":"&#165;","url":"modules/_favware_unescape.html#charsets.__165_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":330,"kind":32,"name":"&#xa5;","url":"modules/_favware_unescape.html#charsets.__xa5_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":331,"kind":32,"name":"&nbsp;","url":"modules/_favware_unescape.html#charsets._nbsp_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":332,"kind":32,"name":"&#160;","url":"modules/_favware_unescape.html#charsets.__160_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":333,"kind":32,"name":"&#xa0;","url":"modules/_favware_unescape.html#charsets.__xa0_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":334,"kind":64,"name":"unescape","url":"modules/_favware_unescape.html#unescape","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/unescape"},{"id":335,"kind":128,"name":"YamlReaderError","url":"classes/_favware_yamlreader.yamlreadererror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":336,"kind":512,"name":"constructor","url":"classes/_favware_yamlreader.yamlreadererror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-private","parent":"@favware/yamlreader.YamlReaderError"},{"id":337,"kind":1024,"name":"name","url":"classes/_favware_yamlreader.yamlreadererror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/yamlreader.YamlReaderError"},{"id":338,"kind":1024,"name":"message","url":"classes/_favware_yamlreader.yamlreadererror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/yamlreader.YamlReaderError"},{"id":339,"kind":1024,"name":"stack","url":"classes/_favware_yamlreader.yamlreadererror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/yamlreader.YamlReaderError"},{"id":340,"kind":1024,"name":"Error","url":"classes/_favware_yamlreader.yamlreadererror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/yamlreader.YamlReaderError"},{"id":341,"kind":4194304,"name":"YamlReaderOptions","url":"modules/_favware_yamlreader.html#yamlreaderoptions","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":342,"kind":64,"name":"readYaml","url":"modules/_favware_yamlreader.html#readyaml","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":343,"kind":4194304,"name":"PrismVideoFormat","url":"modules/_favware_ytdl_prismplayer.html#prismvideoformat","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"@favware/ytdl-prismplayer"},{"id":344,"kind":64,"name":"play","url":"modules/_favware_ytdl_prismplayer.html#play","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/ytdl-prismplayer"},{"id":345,"kind":256,"name":"Chars","url":"interfaces/_favware_zalgo.chars.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":346,"kind":1024,"name":"all","url":"interfaces/_favware_zalgo.chars.html#all","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":347,"kind":1024,"name":"up","url":"interfaces/_favware_zalgo.chars.html#up","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":348,"kind":1024,"name":"middle","url":"interfaces/_favware_zalgo.chars.html#middle","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":349,"kind":1024,"name":"down","url":"interfaces/_favware_zalgo.chars.html#down","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":350,"kind":1024,"name":"pattern","url":"interfaces/_favware_zalgo.chars.html#pattern","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":351,"kind":2097152,"name":"chars","url":"modules/_favware_zalgo.html#chars-1","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":352,"kind":32,"name":"up","url":"modules/_favware_zalgo.html#chars-1.up","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/zalgo.chars"},{"id":353,"kind":32,"name":"middle","url":"modules/_favware_zalgo.html#chars-1.middle","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/zalgo.chars"},{"id":354,"kind":32,"name":"down","url":"modules/_favware_zalgo.html#chars-1.down","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/zalgo.chars"},{"id":355,"kind":256,"name":"ZalgoOptions","url":"interfaces/_favware_zalgo.zalgooptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":356,"kind":1024,"name":"up","url":"interfaces/_favware_zalgo.zalgooptions.html#up","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":357,"kind":1024,"name":"middle","url":"interfaces/_favware_zalgo.zalgooptions.html#middle","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":358,"kind":1024,"name":"down","url":"interfaces/_favware_zalgo.zalgooptions.html#down","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":359,"kind":1024,"name":"size","url":"interfaces/_favware_zalgo.zalgooptions.html#size","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":360,"kind":64,"name":"zalgo","url":"modules/_favware_zalgo.html#zalgo","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":361,"kind":64,"name":"banish","url":"modules/_favware_zalgo.html#banish","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":362,"kind":2,"name":"@favware/converter","url":"modules/_favware_converter.html","classes":"tsd-kind-module"},{"id":363,"kind":2,"name":"@favware/crypto","url":"modules/_favware_crypto.html","classes":"tsd-kind-module"},{"id":364,"kind":2,"name":"@favware/eslint-config","url":"modules/_favware_eslint_config.html","classes":"tsd-kind-module"},{"id":365,"kind":2,"name":"@favware/milky-tslint","url":"modules/_favware_milky_tslint.html","classes":"tsd-kind-module"},{"id":366,"kind":2,"name":"@favware/querystring","url":"modules/_favware_querystring.html","classes":"tsd-kind-module"},{"id":367,"kind":2,"name":"@favware/unescape","url":"modules/_favware_unescape.html","classes":"tsd-kind-module"},{"id":368,"kind":2,"name":"@favware/yamlreader","url":"modules/_favware_yamlreader.html","classes":"tsd-kind-module"},{"id":369,"kind":2,"name":"@favware/ytdl-prismplayer","url":"modules/_favware_ytdl_prismplayer.html","classes":"tsd-kind-module"},{"id":370,"kind":2,"name":"@favware/zalgo","url":"modules/_favware_zalgo.html","classes":"tsd-kind-module"}]};
+            typedoc.search.data = {"kinds":{"2":"Module","4":"Enumeration","16":"Enumeration member","32":"Variable","64":"Function","128":"Class","256":"Interface","512":"Constructor","1024":"Property","2048":"Method","65536":"Type literal","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":256,"name":"MeasurementSystemData","url":"interfaces/_favware_converter.measurementsystemdata.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":1,"kind":1024,"name":"default","url":"interfaces/_favware_converter.measurementsystemdata.html#default","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.MeasurementSystemData"},{"id":2,"kind":1024,"name":"ratio","url":"interfaces/_favware_converter.measurementsystemdata.html#ratio","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.MeasurementSystemData"},{"id":3,"kind":2048,"name":"transform","url":"interfaces/_favware_converter.measurementsystemdata.html#transform","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"@favware/converter.MeasurementSystemData"},{"id":4,"kind":256,"name":"UnitDefinition","url":"interfaces/_favware_converter.unitdefinition.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":5,"kind":1024,"name":"name","url":"interfaces/_favware_converter.unitdefinition.html#name","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":6,"kind":1024,"name":"primary","url":"interfaces/_favware_converter.unitdefinition.html#primary","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":7,"kind":1024,"name":"secondary","url":"interfaces/_favware_converter.unitdefinition.html#secondary","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":8,"kind":1024,"name":"data","url":"interfaces/_favware_converter.unitdefinition.html#data","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitDefinition"},{"id":9,"kind":256,"name":"UnitData","url":"interfaces/_favware_converter.unitdata.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":10,"kind":1024,"name":"id","url":"interfaces/_favware_converter.unitdata.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":11,"kind":1024,"name":"system","url":"interfaces/_favware_converter.unitdata.html#system","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":12,"kind":1024,"name":"multiplier","url":"interfaces/_favware_converter.unitdata.html#multiplier","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":13,"kind":1024,"name":"valueShift","url":"interfaces/_favware_converter.unitdata.html#valueshift","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.UnitData"},{"id":14,"kind":256,"name":"ConvertOptions","url":"interfaces/_favware_converter.convertoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/converter"},{"id":15,"kind":1024,"name":"precision","url":"interfaces/_favware_converter.convertoptions.html#precision","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/converter.ConvertOptions"},{"id":16,"kind":4,"name":"System","url":"enums/_favware_converter.system.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/converter"},{"id":17,"kind":16,"name":"PRIMARY","url":"enums/_favware_converter.system.html#primary","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/converter.System"},{"id":18,"kind":16,"name":"SECONDARY","url":"enums/_favware_converter.system.html#secondary","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/converter.System"},{"id":19,"kind":128,"name":"ConverterError","url":"classes/_favware_converter.convertererror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/converter"},{"id":20,"kind":512,"name":"constructor","url":"classes/_favware_converter.convertererror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"@favware/converter.ConverterError"},{"id":21,"kind":1024,"name":"name","url":"classes/_favware_converter.convertererror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/converter.ConverterError"},{"id":22,"kind":1024,"name":"message","url":"classes/_favware_converter.convertererror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/converter.ConverterError"},{"id":23,"kind":1024,"name":"stack","url":"classes/_favware_converter.convertererror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/converter.ConverterError"},{"id":24,"kind":1024,"name":"Error","url":"classes/_favware_converter.convertererror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/converter.ConverterError"},{"id":25,"kind":32,"name":"definitions","url":"modules/_favware_converter.html#definitions","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/converter"},{"id":26,"kind":64,"name":"roundNumber","url":"modules/_favware_converter.html#roundnumber","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/converter"},{"id":27,"kind":64,"name":"convert","url":"modules/_favware_converter.html#convert","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/converter"},{"id":28,"kind":4,"name":"Radix","url":"enums/_favware_crypto.radix.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/crypto"},{"id":29,"kind":16,"name":"base16","url":"enums/_favware_crypto.radix.html#base16","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":30,"kind":16,"name":"base32","url":"enums/_favware_crypto.radix.html#base32","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":31,"kind":16,"name":"base36","url":"enums/_favware_crypto.radix.html#base36","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":32,"kind":16,"name":"base58","url":"enums/_favware_crypto.radix.html#base58","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":33,"kind":16,"name":"base64","url":"enums/_favware_crypto.radix.html#base64","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/crypto.Radix"},{"id":34,"kind":128,"name":"CryptoError","url":"classes/_favware_crypto.cryptoerror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/crypto"},{"id":35,"kind":512,"name":"constructor","url":"classes/_favware_crypto.cryptoerror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"@favware/crypto.CryptoError"},{"id":36,"kind":1024,"name":"name","url":"classes/_favware_crypto.cryptoerror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/crypto.CryptoError"},{"id":37,"kind":1024,"name":"message","url":"classes/_favware_crypto.cryptoerror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/crypto.CryptoError"},{"id":38,"kind":1024,"name":"stack","url":"classes/_favware_crypto.cryptoerror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/crypto.CryptoError"},{"id":39,"kind":1024,"name":"Error","url":"classes/_favware_crypto.cryptoerror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/crypto.CryptoError"},{"id":40,"kind":32,"name":"DEFAULT_BITS","url":"modules/_favware_crypto.html#default_bits","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/crypto"},{"id":41,"kind":32,"name":"DEFAULT_RADIX","url":"modules/_favware_crypto.html#default_radix","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/crypto"},{"id":42,"kind":32,"name":"CHARACTERS","url":"modules/_favware_crypto.html#characters","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/crypto"},{"id":43,"kind":64,"name":"crypto","url":"modules/_favware_crypto.html#crypto","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/crypto"},{"id":44,"kind":2097152,"name":"config","url":"modules/_favware_eslint_config.html#config","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/eslint-config"},{"id":45,"kind":32,"name":"root","url":"modules/_favware_eslint_config.html#config.root","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":46,"kind":32,"name":"parser","url":"modules/_favware_eslint_config.html#config.parser","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":47,"kind":2097152,"name":"parserOptions","url":"modules/_favware_eslint_config.html#config.parseroptions","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":48,"kind":32,"name":"project","url":"modules/_favware_eslint_config.html#config.parseroptions.project","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":49,"kind":32,"name":"jsx","url":"modules/_favware_eslint_config.html#config.parseroptions.jsx","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":50,"kind":32,"name":"sourceType","url":"modules/_favware_eslint_config.html#config.parseroptions.sourcetype","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":51,"kind":32,"name":"ecmaVersion","url":"modules/_favware_eslint_config.html#config.parseroptions.ecmaversion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":52,"kind":2097152,"name":"ecmaFeatures","url":"modules/_favware_eslint_config.html#config.parseroptions.ecmafeatures","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions"},{"id":53,"kind":32,"name":"jsx","url":"modules/_favware_eslint_config.html#config.parseroptions.ecmafeatures.jsx-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.parserOptions.ecmaFeatures"},{"id":54,"kind":32,"name":"extends","url":"modules/_favware_eslint_config.html#config.extends","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":55,"kind":2097152,"name":"env","url":"modules/_favware_eslint_config.html#config.env","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":56,"kind":32,"name":"node","url":"modules/_favware_eslint_config.html#config.env.node","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":57,"kind":32,"name":"es6","url":"modules/_favware_eslint_config.html#config.env.es6","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":58,"kind":32,"name":"jest","url":"modules/_favware_eslint_config.html#config.env.jest","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":59,"kind":32,"name":"browser","url":"modules/_favware_eslint_config.html#config.env.browser","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":60,"kind":32,"name":"commonjs","url":"modules/_favware_eslint_config.html#config.env.commonjs","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.env"},{"id":61,"kind":2097152,"name":"rules","url":"modules/_favware_eslint_config.html#config.rules","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config"},{"id":62,"kind":32,"name":"@typescript-eslint/indent","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_indent","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":63,"kind":32,"name":"@typescript-eslint/no-non-null-assertion","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_non_null_assertion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":64,"kind":32,"name":"@typescript-eslint/explicit-function-return-type","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_explicit_function_return_type","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":65,"kind":32,"name":"@typescript-eslint/prefer-interface","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_prefer_interface","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":66,"kind":32,"name":"@typescript-eslint/consistent-type-definitions","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_consistent_type_definitions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":67,"kind":32,"name":"@typescript-eslint/interface-name-prefix","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_interface_name_prefix","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":68,"kind":32,"name":"@typescript-eslint/await-thenable","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_await_thenable","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":69,"kind":32,"name":"@typescript-eslint/ban-types","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_ban_types","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":70,"kind":32,"name":"@typescript-eslint/func-call-spacing","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_func_call_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":71,"kind":32,"name":"@typescript-eslint/member-ordering","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_member_ordering","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":72,"kind":32,"name":"@typescript-eslint/no-empty-function","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_empty_function","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":73,"kind":32,"name":"@typescript-eslint/no-extra-parens","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_extra_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":74,"kind":32,"name":"@typescript-eslint/no-extraneous-class","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_extraneous_class","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":75,"kind":32,"name":"@typescript-eslint/no-for-in-array","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_for_in_array","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":76,"kind":32,"name":"@typescript-eslint/no-require-imports","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_require_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":77,"kind":32,"name":"@typescript-eslint/no-this-alias","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_this_alias","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":78,"kind":32,"name":"@typescript-eslint/no-unnecessary-qualifier","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_unnecessary_qualifier","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":79,"kind":32,"name":"@typescript-eslint/no-unnecessary-type-assertion","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_unnecessary_type_assertion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":80,"kind":32,"name":"@typescript-eslint/no-useless-constructor","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_no_useless_constructor","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":81,"kind":32,"name":"@typescript-eslint/prefer-for-of","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_prefer_for_of","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":82,"kind":32,"name":"@typescript-eslint/prefer-includes","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_prefer_includes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":83,"kind":32,"name":"@typescript-eslint/promise-function-async","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_promise_function_async","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":84,"kind":32,"name":"@typescript-eslint/require-array-sort-compare","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_require_array_sort_compare","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":85,"kind":32,"name":"@typescript-eslint/restrict-plus-operands","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_restrict_plus_operands","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":86,"kind":32,"name":"@typescript-eslint/semi","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_semi","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":87,"kind":32,"name":"@typescript-eslint/unbound-method","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_unbound_method","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":88,"kind":32,"name":"@typescript-eslint/unified-signatures","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_unified_signatures","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":89,"kind":32,"name":"@typescript-eslint/camelcase","url":"modules/_favware_eslint_config.html#config.rules._typescript_eslint_camelcase","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":90,"kind":32,"name":"jest/consistent-test-it","url":"modules/_favware_eslint_config.html#config.rules.jest_consistent_test_it","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":91,"kind":32,"name":"jest/no-hooks","url":"modules/_favware_eslint_config.html#config.rules.jest_no_hooks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":92,"kind":32,"name":"jest/lowercase-name","url":"modules/_favware_eslint_config.html#config.rules.jest_lowercase_name","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":93,"kind":32,"name":"jest/no-export","url":"modules/_favware_eslint_config.html#config.rules.jest_no_export","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":94,"kind":32,"name":"jest/prefer-expect-assertions","url":"modules/_favware_eslint_config.html#config.rules.jest_prefer_expect_assertions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":95,"kind":32,"name":"jest/no-if","url":"modules/_favware_eslint_config.html#config.rules.jest_no_if","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":96,"kind":32,"name":"jest/prefer-inline-snapshots","url":"modules/_favware_eslint_config.html#config.rules.jest_prefer_inline_snapshots","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":97,"kind":32,"name":"jest/expect-expect","url":"modules/_favware_eslint_config.html#config.rules.jest_expect_expect","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":98,"kind":32,"name":"jest/require-top-level-describe","url":"modules/_favware_eslint_config.html#config.rules.jest_require_top_level_describe","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":99,"kind":32,"name":"no-console","url":"modules/_favware_eslint_config.html#config.rules.no_console","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":100,"kind":32,"name":"no-template-curly-in-string","url":"modules/_favware_eslint_config.html#config.rules.no_template_curly_in_string","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":101,"kind":32,"name":"require-atomic-updates","url":"modules/_favware_eslint_config.html#config.rules.require_atomic_updates","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":102,"kind":32,"name":"accessor-pairs","url":"modules/_favware_eslint_config.html#config.rules.accessor_pairs","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":103,"kind":32,"name":"array-callback-return","url":"modules/_favware_eslint_config.html#config.rules.array_callback_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":104,"kind":32,"name":"block-scoped-var","url":"modules/_favware_eslint_config.html#config.rules.block_scoped_var","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":105,"kind":32,"name":"consistent-return","url":"modules/_favware_eslint_config.html#config.rules.consistent_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":106,"kind":32,"name":"curly","url":"modules/_favware_eslint_config.html#config.rules.curly","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":107,"kind":32,"name":"default-case","url":"modules/_favware_eslint_config.html#config.rules.default_case","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":108,"kind":32,"name":"dot-location","url":"modules/_favware_eslint_config.html#config.rules.dot_location","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":109,"kind":32,"name":"eqeqeq","url":"modules/_favware_eslint_config.html#config.rules.eqeqeq","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":110,"kind":32,"name":"max-classes-per-file","url":"modules/_favware_eslint_config.html#config.rules.max_classes_per_file","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":111,"kind":32,"name":"no-alert","url":"modules/_favware_eslint_config.html#config.rules.no_alert","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":112,"kind":32,"name":"no-caller","url":"modules/_favware_eslint_config.html#config.rules.no_caller","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":113,"kind":32,"name":"no-div-regex","url":"modules/_favware_eslint_config.html#config.rules.no_div_regex","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":114,"kind":32,"name":"no-else-return","url":"modules/_favware_eslint_config.html#config.rules.no_else_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":115,"kind":32,"name":"no-eq-null","url":"modules/_favware_eslint_config.html#config.rules.no_eq_null","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":116,"kind":32,"name":"no-eval","url":"modules/_favware_eslint_config.html#config.rules.no_eval","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":117,"kind":32,"name":"no-extend-native","url":"modules/_favware_eslint_config.html#config.rules.no_extend_native","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":118,"kind":32,"name":"no-extra-bind","url":"modules/_favware_eslint_config.html#config.rules.no_extra_bind","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":119,"kind":32,"name":"no-extra-label","url":"modules/_favware_eslint_config.html#config.rules.no_extra_label","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":120,"kind":32,"name":"no-floating-decimal","url":"modules/_favware_eslint_config.html#config.rules.no_floating_decimal","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":121,"kind":32,"name":"no-implicit-coercion","url":"modules/_favware_eslint_config.html#config.rules.no_implicit_coercion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":122,"kind":32,"name":"no-implicit-globals","url":"modules/_favware_eslint_config.html#config.rules.no_implicit_globals","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":123,"kind":32,"name":"no-implied-eval","url":"modules/_favware_eslint_config.html#config.rules.no_implied_eval","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":124,"kind":32,"name":"no-invalid-this","url":"modules/_favware_eslint_config.html#config.rules.no_invalid_this","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":125,"kind":32,"name":"no-iterator","url":"modules/_favware_eslint_config.html#config.rules.no_iterator","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":126,"kind":32,"name":"no-lone-blocks","url":"modules/_favware_eslint_config.html#config.rules.no_lone_blocks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":127,"kind":32,"name":"no-loop-func","url":"modules/_favware_eslint_config.html#config.rules.no_loop_func","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":128,"kind":32,"name":"no-multi-spaces","url":"modules/_favware_eslint_config.html#config.rules.no_multi_spaces","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":129,"kind":32,"name":"no-multi-str","url":"modules/_favware_eslint_config.html#config.rules.no_multi_str","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":130,"kind":32,"name":"no-new","url":"modules/_favware_eslint_config.html#config.rules.no_new","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":131,"kind":32,"name":"no-new-func","url":"modules/_favware_eslint_config.html#config.rules.no_new_func","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":132,"kind":32,"name":"no-new-wrappers","url":"modules/_favware_eslint_config.html#config.rules.no_new_wrappers","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":133,"kind":32,"name":"no-octal-escape","url":"modules/_favware_eslint_config.html#config.rules.no_octal_escape","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":134,"kind":32,"name":"no-proto","url":"modules/_favware_eslint_config.html#config.rules.no_proto","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":135,"kind":32,"name":"no-prototype-builtins","url":"modules/_favware_eslint_config.html#config.rules.no_prototype_builtins","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":136,"kind":32,"name":"no-return-assign","url":"modules/_favware_eslint_config.html#config.rules.no_return_assign","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":137,"kind":32,"name":"no-return-await","url":"modules/_favware_eslint_config.html#config.rules.no_return_await","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":138,"kind":32,"name":"no-script-url","url":"modules/_favware_eslint_config.html#config.rules.no_script_url","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":139,"kind":32,"name":"no-self-compare","url":"modules/_favware_eslint_config.html#config.rules.no_self_compare","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":140,"kind":32,"name":"no-sequences","url":"modules/_favware_eslint_config.html#config.rules.no_sequences","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":141,"kind":32,"name":"no-throw-literal","url":"modules/_favware_eslint_config.html#config.rules.no_throw_literal","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":142,"kind":32,"name":"no-unmodified-loop-condition","url":"modules/_favware_eslint_config.html#config.rules.no_unmodified_loop_condition","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":143,"kind":32,"name":"no-unused-expressions","url":"modules/_favware_eslint_config.html#config.rules.no_unused_expressions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":144,"kind":32,"name":"no-useless-call","url":"modules/_favware_eslint_config.html#config.rules.no_useless_call","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":145,"kind":32,"name":"no-useless-concat","url":"modules/_favware_eslint_config.html#config.rules.no_useless_concat","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":146,"kind":32,"name":"no-useless-return","url":"modules/_favware_eslint_config.html#config.rules.no_useless_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":147,"kind":32,"name":"no-void","url":"modules/_favware_eslint_config.html#config.rules.no_void","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":148,"kind":32,"name":"no-warning-comments","url":"modules/_favware_eslint_config.html#config.rules.no_warning_comments","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":149,"kind":32,"name":"prefer-promise-reject-errors","url":"modules/_favware_eslint_config.html#config.rules.prefer_promise_reject_errors","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":150,"kind":32,"name":"radix","url":"modules/_favware_eslint_config.html#config.rules.radix","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":151,"kind":32,"name":"wrap-iife","url":"modules/_favware_eslint_config.html#config.rules.wrap_iife","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":152,"kind":32,"name":"yoda","url":"modules/_favware_eslint_config.html#config.rules.yoda","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":153,"kind":32,"name":"strict","url":"modules/_favware_eslint_config.html#config.rules.strict","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":154,"kind":32,"name":"no-label-var","url":"modules/_favware_eslint_config.html#config.rules.no_label_var","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":155,"kind":32,"name":"no-shadow","url":"modules/_favware_eslint_config.html#config.rules.no_shadow","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":156,"kind":32,"name":"no-undef-init","url":"modules/_favware_eslint_config.html#config.rules.no_undef_init","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":157,"kind":32,"name":"callback-return","url":"modules/_favware_eslint_config.html#config.rules.callback_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":158,"kind":32,"name":"global-require","url":"modules/_favware_eslint_config.html#config.rules.global_require","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":159,"kind":32,"name":"handle-callback-err","url":"modules/_favware_eslint_config.html#config.rules.handle_callback_err","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":160,"kind":32,"name":"no-buffer-constructor","url":"modules/_favware_eslint_config.html#config.rules.no_buffer_constructor","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":161,"kind":32,"name":"no-mixed-requires","url":"modules/_favware_eslint_config.html#config.rules.no_mixed_requires","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":162,"kind":32,"name":"no-new-require","url":"modules/_favware_eslint_config.html#config.rules.no_new_require","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":163,"kind":32,"name":"no-path-concat","url":"modules/_favware_eslint_config.html#config.rules.no_path_concat","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":164,"kind":32,"name":"array-bracket-spacing","url":"modules/_favware_eslint_config.html#config.rules.array_bracket_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":165,"kind":32,"name":"block-spacing","url":"modules/_favware_eslint_config.html#config.rules.block_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":166,"kind":32,"name":"brace-style","url":"modules/_favware_eslint_config.html#config.rules.brace_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":167,"kind":32,"name":"capitalized-comments","url":"modules/_favware_eslint_config.html#config.rules.capitalized_comments","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":168,"kind":32,"name":"comma-dangle","url":"modules/_favware_eslint_config.html#config.rules.comma_dangle","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":169,"kind":32,"name":"comma-spacing","url":"modules/_favware_eslint_config.html#config.rules.comma_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":170,"kind":32,"name":"comma-style","url":"modules/_favware_eslint_config.html#config.rules.comma_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":171,"kind":32,"name":"computed-property-spacing","url":"modules/_favware_eslint_config.html#config.rules.computed_property_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":172,"kind":32,"name":"eol-last","url":"modules/_favware_eslint_config.html#config.rules.eol_last","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":173,"kind":32,"name":"func-name-matching","url":"modules/_favware_eslint_config.html#config.rules.func_name_matching","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":174,"kind":32,"name":"func-names","url":"modules/_favware_eslint_config.html#config.rules.func_names","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":175,"kind":32,"name":"func-style","url":"modules/_favware_eslint_config.html#config.rules.func_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":176,"kind":32,"name":"implicit-arrow-linebreak","url":"modules/_favware_eslint_config.html#config.rules.implicit_arrow_linebreak","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":177,"kind":32,"name":"jsx-quotes","url":"modules/_favware_eslint_config.html#config.rules.jsx_quotes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":178,"kind":32,"name":"key-spacing","url":"modules/_favware_eslint_config.html#config.rules.key_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":179,"kind":32,"name":"keyword-spacing","url":"modules/_favware_eslint_config.html#config.rules.keyword_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":180,"kind":32,"name":"linebreak-style","url":"modules/_favware_eslint_config.html#config.rules.linebreak_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":181,"kind":32,"name":"lines-between-class-members","url":"modules/_favware_eslint_config.html#config.rules.lines_between_class_members","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":182,"kind":32,"name":"max-len","url":"modules/_favware_eslint_config.html#config.rules.max_len","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":183,"kind":32,"name":"max-statements-per-line","url":"modules/_favware_eslint_config.html#config.rules.max_statements_per_line","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":184,"kind":32,"name":"new-cap","url":"modules/_favware_eslint_config.html#config.rules.new_cap","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":185,"kind":32,"name":"new-parens","url":"modules/_favware_eslint_config.html#config.rules.new_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":186,"kind":32,"name":"no-array-constructor","url":"modules/_favware_eslint_config.html#config.rules.no_array_constructor","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":187,"kind":32,"name":"no-lonely-if","url":"modules/_favware_eslint_config.html#config.rules.no_lonely_if","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":188,"kind":32,"name":"no-mixed-operators","url":"modules/_favware_eslint_config.html#config.rules.no_mixed_operators","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":189,"kind":32,"name":"no-multi-assign","url":"modules/_favware_eslint_config.html#config.rules.no_multi_assign","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":190,"kind":32,"name":"no-multiple-empty-lines","url":"modules/_favware_eslint_config.html#config.rules.no_multiple_empty_lines","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":191,"kind":32,"name":"no-nested-ternary","url":"modules/_favware_eslint_config.html#config.rules.no_nested_ternary","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":192,"kind":32,"name":"no-new-object","url":"modules/_favware_eslint_config.html#config.rules.no_new_object","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":193,"kind":32,"name":"no-tabs","url":"modules/_favware_eslint_config.html#config.rules.no_tabs","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":194,"kind":32,"name":"no-trailing-spaces","url":"modules/_favware_eslint_config.html#config.rules.no_trailing_spaces","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":195,"kind":32,"name":"no-underscore-dangle","url":"modules/_favware_eslint_config.html#config.rules.no_underscore_dangle","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":196,"kind":32,"name":"no-unneeded-ternary","url":"modules/_favware_eslint_config.html#config.rules.no_unneeded_ternary","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":197,"kind":32,"name":"no-whitespace-before-property","url":"modules/_favware_eslint_config.html#config.rules.no_whitespace_before_property","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":198,"kind":32,"name":"nonblock-statement-body-position","url":"modules/_favware_eslint_config.html#config.rules.nonblock_statement_body_position","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":199,"kind":32,"name":"object-curly-newline","url":"modules/_favware_eslint_config.html#config.rules.object_curly_newline","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":200,"kind":32,"name":"one-var","url":"modules/_favware_eslint_config.html#config.rules.one_var","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":201,"kind":32,"name":"one-var-declaration-per-line","url":"modules/_favware_eslint_config.html#config.rules.one_var_declaration_per_line","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":202,"kind":32,"name":"operator-linebreak","url":"modules/_favware_eslint_config.html#config.rules.operator_linebreak","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":203,"kind":32,"name":"padded-blocks","url":"modules/_favware_eslint_config.html#config.rules.padded_blocks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":204,"kind":32,"name":"padding-line-between-statements","url":"modules/_favware_eslint_config.html#config.rules.padding_line_between_statements","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":205,"kind":32,"name":"prefer-object-spread","url":"modules/_favware_eslint_config.html#config.rules.prefer_object_spread","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":206,"kind":32,"name":"quote-props","url":"modules/_favware_eslint_config.html#config.rules.quote_props","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":207,"kind":32,"name":"quotes","url":"modules/_favware_eslint_config.html#config.rules.quotes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":208,"kind":32,"name":"semi-spacing","url":"modules/_favware_eslint_config.html#config.rules.semi_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":209,"kind":32,"name":"semi-style","url":"modules/_favware_eslint_config.html#config.rules.semi_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":210,"kind":32,"name":"sort-vars","url":"modules/_favware_eslint_config.html#config.rules.sort_vars","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":211,"kind":32,"name":"space-before-blocks","url":"modules/_favware_eslint_config.html#config.rules.space_before_blocks","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":212,"kind":32,"name":"space-before-function-paren","url":"modules/_favware_eslint_config.html#config.rules.space_before_function_paren","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":213,"kind":32,"name":"space-in-parens","url":"modules/_favware_eslint_config.html#config.rules.space_in_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":214,"kind":32,"name":"space-infix-ops","url":"modules/_favware_eslint_config.html#config.rules.space_infix_ops","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":215,"kind":32,"name":"space-unary-ops","url":"modules/_favware_eslint_config.html#config.rules.space_unary_ops","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":216,"kind":32,"name":"switch-colon-spacing","url":"modules/_favware_eslint_config.html#config.rules.switch_colon_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":217,"kind":32,"name":"template-tag-spacing","url":"modules/_favware_eslint_config.html#config.rules.template_tag_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":218,"kind":32,"name":"no-useless-computed-key","url":"modules/_favware_eslint_config.html#config.rules.no_useless_computed_key","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":219,"kind":32,"name":"arrow-parens","url":"modules/_favware_eslint_config.html#config.rules.arrow_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":220,"kind":32,"name":"arrow-spacing","url":"modules/_favware_eslint_config.html#config.rules.arrow_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":221,"kind":32,"name":"generator-star-spacing","url":"modules/_favware_eslint_config.html#config.rules.generator_star_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":222,"kind":32,"name":"no-duplicate-imports","url":"modules/_favware_eslint_config.html#config.rules.no_duplicate_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":223,"kind":32,"name":"no-useless-rename","url":"modules/_favware_eslint_config.html#config.rules.no_useless_rename","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":224,"kind":32,"name":"prefer-arrow-callback","url":"modules/_favware_eslint_config.html#config.rules.prefer_arrow_callback","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":225,"kind":32,"name":"prefer-numeric-literals","url":"modules/_favware_eslint_config.html#config.rules.prefer_numeric_literals","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":226,"kind":32,"name":"prefer-template","url":"modules/_favware_eslint_config.html#config.rules.prefer_template","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":227,"kind":32,"name":"rest-spread-spacing","url":"modules/_favware_eslint_config.html#config.rules.rest_spread_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":228,"kind":32,"name":"template-curly-spacing","url":"modules/_favware_eslint_config.html#config.rules.template_curly_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":229,"kind":32,"name":"yield-star-spacing","url":"modules/_favware_eslint_config.html#config.rules.yield_star_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/eslint-config.config.rules"},{"id":230,"kind":256,"name":"PluginOptions","url":"interfaces/_favware_milky_tslint.pluginoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":231,"kind":1024,"name":"configuration","url":"interfaces/_favware_milky_tslint.pluginoptions.html#configuration","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":232,"kind":1024,"name":"fix","url":"interfaces/_favware_milky_tslint.pluginoptions.html#fix","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":233,"kind":1024,"name":"quiet","url":"interfaces/_favware_milky_tslint.pluginoptions.html#quiet","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":234,"kind":1024,"name":"formatter","url":"interfaces/_favware_milky_tslint.pluginoptions.html#formatter","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":235,"kind":1024,"name":"formattersDirectory","url":"interfaces/_favware_milky_tslint.pluginoptions.html#formattersdirectory","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":236,"kind":1024,"name":"rulesDirectory","url":"interfaces/_favware_milky_tslint.pluginoptions.html#rulesdirectory","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":237,"kind":1024,"name":"tslint","url":"interfaces/_favware_milky_tslint.pluginoptions.html#tslint","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":238,"kind":1024,"name":"program","url":"interfaces/_favware_milky_tslint.pluginoptions.html#program","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.PluginOptions"},{"id":239,"kind":256,"name":"ReportOptions","url":"interfaces/_favware_milky_tslint.reportoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":240,"kind":1024,"name":"emitError","url":"interfaces/_favware_milky_tslint.reportoptions.html#emiterror","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":241,"kind":1024,"name":"reportLimit","url":"interfaces/_favware_milky_tslint.reportoptions.html#reportlimit","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":242,"kind":1024,"name":"summarizeFailureOutput","url":"interfaces/_favware_milky_tslint.reportoptions.html#summarizefailureoutput","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":243,"kind":1024,"name":"allowWarnings","url":"interfaces/_favware_milky_tslint.reportoptions.html#allowwarnings","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.ReportOptions"},{"id":244,"kind":256,"name":"TSLintFile","url":"interfaces/_favware_milky_tslint.tslintfile.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":245,"kind":1024,"name":"tslint","url":"interfaces/_favware_milky_tslint.tslintfile.html#tslint","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":246,"kind":1024,"name":"path","url":"interfaces/_favware_milky_tslint.tslintfile.html#path","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":247,"kind":1024,"name":"relative","url":"interfaces/_favware_milky_tslint.tslintfile.html#relative","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":248,"kind":1024,"name":"contents","url":"interfaces/_favware_milky_tslint.tslintfile.html#contents","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":249,"kind":2048,"name":"isStream","url":"interfaces/_favware_milky_tslint.tslintfile.html#isstream","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":250,"kind":2048,"name":"isNull","url":"interfaces/_favware_milky_tslint.tslintfile.html#isnull","classes":"tsd-kind-method tsd-parent-kind-interface","parent":"@favware/milky-tslint.TSLintFile"},{"id":251,"kind":4,"name":"LogLevels","url":"enums/_favware_milky_tslint.loglevels.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":252,"kind":16,"name":"ERROR","url":"enums/_favware_milky_tslint.loglevels.html#error","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LogLevels"},{"id":253,"kind":16,"name":"INFO","url":"enums/_favware_milky_tslint.loglevels.html#info","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LogLevels"},{"id":254,"kind":16,"name":"WARN","url":"enums/_favware_milky_tslint.loglevels.html#warn","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LogLevels"},{"id":255,"kind":4,"name":"LintFormatters","url":"enums/_favware_milky_tslint.lintformatters.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":256,"kind":16,"name":"CHECKSTYLE","url":"enums/_favware_milky_tslint.lintformatters.html#checkstyle","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":257,"kind":16,"name":"CODEFRAME","url":"enums/_favware_milky_tslint.lintformatters.html#codeframe","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":258,"kind":16,"name":"FILESLIST","url":"enums/_favware_milky_tslint.lintformatters.html#fileslist","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":259,"kind":16,"name":"JSON","url":"enums/_favware_milky_tslint.lintformatters.html#json","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":260,"kind":16,"name":"JUNIT","url":"enums/_favware_milky_tslint.lintformatters.html#junit","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":261,"kind":16,"name":"MSBUILD","url":"enums/_favware_milky_tslint.lintformatters.html#msbuild","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":262,"kind":16,"name":"PMD","url":"enums/_favware_milky_tslint.lintformatters.html#pmd","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":263,"kind":16,"name":"PROSE","url":"enums/_favware_milky_tslint.lintformatters.html#prose","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":264,"kind":16,"name":"STYLISH","url":"enums/_favware_milky_tslint.lintformatters.html#stylish","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":265,"kind":16,"name":"TAP","url":"enums/_favware_milky_tslint.lintformatters.html#tap","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":266,"kind":16,"name":"VERBOSE","url":"enums/_favware_milky_tslint.lintformatters.html#verbose","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":267,"kind":16,"name":"VSO","url":"enums/_favware_milky_tslint.lintformatters.html#vso","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/milky-tslint.LintFormatters"},{"id":268,"kind":256,"name":"MilkyEvents","url":"interfaces/_favware_milky_tslint.milkyevents.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":269,"kind":1024,"name":"data","url":"interfaces/_favware_milky_tslint.milkyevents.html#data","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.MilkyEvents"},{"id":270,"kind":65536,"name":"__type","url":"interfaces/_favware_milky_tslint.milkyevents.html#data.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"@favware/milky-tslint.MilkyEvents.data"},{"id":271,"kind":1024,"name":"error","url":"interfaces/_favware_milky_tslint.milkyevents.html#error","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.MilkyEvents"},{"id":272,"kind":65536,"name":"__type","url":"interfaces/_favware_milky_tslint.milkyevents.html#error.__type-2","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"@favware/milky-tslint.MilkyEvents.error"},{"id":273,"kind":1024,"name":"end","url":"interfaces/_favware_milky_tslint.milkyevents.html#end","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/milky-tslint.MilkyEvents"},{"id":274,"kind":65536,"name":"__type","url":"interfaces/_favware_milky_tslint.milkyevents.html#end.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"@favware/milky-tslint.MilkyEvents.end"},{"id":275,"kind":32,"name":"eventemitter","url":"modules/_favware_milky_tslint.html#eventemitter","classes":"tsd-kind-variable tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":276,"kind":64,"name":"isString","url":"modules/_favware_milky_tslint.html#isstring","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":277,"kind":64,"name":"milkyLint","url":"modules/_favware_milky_tslint.html#milkylint","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":278,"kind":64,"name":"milkyReport","url":"modules/_favware_milky_tslint.html#milkyreport","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/milky-tslint"},{"id":279,"kind":256,"name":"StringifyOptions","url":"interfaces/_favware_querystring.stringifyoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/querystring"},{"id":280,"kind":1024,"name":"separator","url":"interfaces/_favware_querystring.stringifyoptions.html#separator","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.StringifyOptions"},{"id":281,"kind":1024,"name":"equals","url":"interfaces/_favware_querystring.stringifyoptions.html#equals","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.StringifyOptions"},{"id":282,"kind":1024,"name":"includeQuestion","url":"interfaces/_favware_querystring.stringifyoptions.html#includequestion","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.StringifyOptions"},{"id":283,"kind":256,"name":"ParseOptions","url":"interfaces/_favware_querystring.parseoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/querystring"},{"id":284,"kind":1024,"name":"separator","url":"interfaces/_favware_querystring.parseoptions.html#separator","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.ParseOptions"},{"id":285,"kind":1024,"name":"equals","url":"interfaces/_favware_querystring.parseoptions.html#equals","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/querystring.ParseOptions"},{"id":286,"kind":256,"name":"QuerystringObject","url":"interfaces/_favware_querystring.querystringobject.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/querystring"},{"id":287,"kind":4,"name":"FunctionSource","url":"enums/_favware_querystring.functionsource.html","classes":"tsd-kind-enum tsd-parent-kind-module","parent":"@favware/querystring"},{"id":288,"kind":16,"name":"Parse","url":"enums/_favware_querystring.functionsource.html#parse","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/querystring.FunctionSource"},{"id":289,"kind":16,"name":"Stringify","url":"enums/_favware_querystring.functionsource.html#stringify","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"@favware/querystring.FunctionSource"},{"id":290,"kind":4194304,"name":"QuerystringError","url":"modules/_favware_querystring.html#querystringerror","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"@favware/querystring"},{"id":291,"kind":64,"name":"stringify","url":"modules/_favware_querystring.html#stringify","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/querystring"},{"id":292,"kind":64,"name":"parse","url":"modules/_favware_querystring.html#parse","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/querystring"},{"id":293,"kind":2097152,"name":"config","url":"modules/_favware_tslint_config.html#config","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/tslint-config"},{"id":294,"kind":2097152,"name":"rules","url":"modules/_favware_tslint_config.html#config.rules","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config"},{"id":295,"kind":32,"name":"align","url":"modules/_favware_tslint_config.html#config.rules.align","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":296,"kind":32,"name":"array-type","url":"modules/_favware_tslint_config.html#config.rules.array_type","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":297,"kind":32,"name":"arrow-parens","url":"modules/_favware_tslint_config.html#config.rules.arrow_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":298,"kind":32,"name":"arrow-return-shorthand","url":"modules/_favware_tslint_config.html#config.rules.arrow_return_shorthand","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":299,"kind":32,"name":"await-promise","url":"modules/_favware_tslint_config.html#config.rules.await_promise","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":300,"kind":32,"name":"ban-comma-operator","url":"modules/_favware_tslint_config.html#config.rules.ban_comma_operator","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":301,"kind":32,"name":"binary-expression-operand-order","url":"modules/_favware_tslint_config.html#config.rules.binary_expression_operand_order","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":302,"kind":32,"name":"class-name","url":"modules/_favware_tslint_config.html#config.rules.class_name","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":303,"kind":32,"name":"comment-format","url":"modules/_favware_tslint_config.html#config.rules.comment_format","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":304,"kind":32,"name":"curly","url":"modules/_favware_tslint_config.html#config.rules.curly","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":305,"kind":32,"name":"cyclomatic-complexity","url":"modules/_favware_tslint_config.html#config.rules.cyclomatic_complexity","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":306,"kind":32,"name":"encoding","url":"modules/_favware_tslint_config.html#config.rules.encoding","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":307,"kind":32,"name":"eofline","url":"modules/_favware_tslint_config.html#config.rules.eofline","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":308,"kind":32,"name":"forin","url":"modules/_favware_tslint_config.html#config.rules.forin","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":309,"kind":32,"name":"import-spacing","url":"modules/_favware_tslint_config.html#config.rules.import_spacing","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":310,"kind":32,"name":"indent","url":"modules/_favware_tslint_config.html#config.rules.indent","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":311,"kind":32,"name":"label-position","url":"modules/_favware_tslint_config.html#config.rules.label_position","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":312,"kind":32,"name":"linebreak-style","url":"modules/_favware_tslint_config.html#config.rules.linebreak_style","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":313,"kind":32,"name":"max-classes-per-file","url":"modules/_favware_tslint_config.html#config.rules.max_classes_per_file","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":314,"kind":32,"name":"max-line-length","url":"modules/_favware_tslint_config.html#config.rules.max_line_length","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":315,"kind":32,"name":"member-ordering","url":"modules/_favware_tslint_config.html#config.rules.member_ordering","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":316,"kind":32,"name":"newline-before-return","url":"modules/_favware_tslint_config.html#config.rules.newline_before_return","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":317,"kind":32,"name":"no-angle-bracket-type-assertion","url":"modules/_favware_tslint_config.html#config.rules.no_angle_bracket_type_assertion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":318,"kind":32,"name":"no-arg","url":"modules/_favware_tslint_config.html#config.rules.no_arg","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":319,"kind":32,"name":"no-bitwise","url":"modules/_favware_tslint_config.html#config.rules.no_bitwise","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":320,"kind":32,"name":"no-boolean-literal-compare","url":"modules/_favware_tslint_config.html#config.rules.no_boolean_literal_compare","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":321,"kind":32,"name":"no-conditional-assignment","url":"modules/_favware_tslint_config.html#config.rules.no_conditional_assignment","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":322,"kind":32,"name":"no-consecutive-blank-lines","url":"modules/_favware_tslint_config.html#config.rules.no_consecutive_blank_lines","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":323,"kind":32,"name":"no-console","url":"modules/_favware_tslint_config.html#config.rules.no_console","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":324,"kind":32,"name":"no-construct","url":"modules/_favware_tslint_config.html#config.rules.no_construct","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":325,"kind":32,"name":"no-debugger","url":"modules/_favware_tslint_config.html#config.rules.no_debugger","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":326,"kind":32,"name":"no-duplicate-imports","url":"modules/_favware_tslint_config.html#config.rules.no_duplicate_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":327,"kind":32,"name":"no-duplicate-super","url":"modules/_favware_tslint_config.html#config.rules.no_duplicate_super","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":328,"kind":32,"name":"no-duplicate-switch-case","url":"modules/_favware_tslint_config.html#config.rules.no_duplicate_switch_case","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":329,"kind":32,"name":"no-duplicate-variable","url":"modules/_favware_tslint_config.html#config.rules.no_duplicate_variable","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":330,"kind":32,"name":"no-dynamic-delete","url":"modules/_favware_tslint_config.html#config.rules.no_dynamic_delete","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":331,"kind":32,"name":"no-empty","url":"modules/_favware_tslint_config.html#config.rules.no_empty","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":332,"kind":32,"name":"no-empty-interface","url":"modules/_favware_tslint_config.html#config.rules.no_empty_interface","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":333,"kind":32,"name":"no-eval","url":"modules/_favware_tslint_config.html#config.rules.no_eval","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":334,"kind":32,"name":"no-implicit-dependencies","url":"modules/_favware_tslint_config.html#config.rules.no_implicit_dependencies","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":335,"kind":32,"name":"no-invalid-template-strings","url":"modules/_favware_tslint_config.html#config.rules.no_invalid_template_strings","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":336,"kind":32,"name":"no-invalid-this","url":"modules/_favware_tslint_config.html#config.rules.no_invalid_this","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":337,"kind":32,"name":"no-misused-new","url":"modules/_favware_tslint_config.html#config.rules.no_misused_new","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":338,"kind":32,"name":"no-null-keyword","url":"modules/_favware_tslint_config.html#config.rules.no_null_keyword","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":339,"kind":32,"name":"no-object-literal-type-assertion","url":"modules/_favware_tslint_config.html#config.rules.no_object_literal_type_assertion","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":340,"kind":32,"name":"no-reference","url":"modules/_favware_tslint_config.html#config.rules.no_reference","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":341,"kind":32,"name":"no-require-imports","url":"modules/_favware_tslint_config.html#config.rules.no_require_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":342,"kind":32,"name":"no-return-await","url":"modules/_favware_tslint_config.html#config.rules.no_return_await","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":343,"kind":32,"name":"no-shadowed-variable","url":"modules/_favware_tslint_config.html#config.rules.no_shadowed_variable","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":344,"kind":32,"name":"no-sparse-arrays","url":"modules/_favware_tslint_config.html#config.rules.no_sparse_arrays","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":345,"kind":32,"name":"no-string-literal","url":"modules/_favware_tslint_config.html#config.rules.no_string_literal","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":346,"kind":32,"name":"no-string-throw","url":"modules/_favware_tslint_config.html#config.rules.no_string_throw","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":347,"kind":32,"name":"no-this-assignment","url":"modules/_favware_tslint_config.html#config.rules.no_this_assignment","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":348,"kind":32,"name":"no-trailing-whitespace","url":"modules/_favware_tslint_config.html#config.rules.no_trailing_whitespace","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":349,"kind":32,"name":"no-unsafe-finally","url":"modules/_favware_tslint_config.html#config.rules.no_unsafe_finally","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":350,"kind":32,"name":"no-unused-expression","url":"modules/_favware_tslint_config.html#config.rules.no_unused_expression","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":351,"kind":32,"name":"no-var-keyword","url":"modules/_favware_tslint_config.html#config.rules.no_var_keyword","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":352,"kind":32,"name":"no-var-requires","url":"modules/_favware_tslint_config.html#config.rules.no_var_requires","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":353,"kind":32,"name":"number-literal-format","url":"modules/_favware_tslint_config.html#config.rules.number_literal_format","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":354,"kind":32,"name":"object-literal-key-quotes","url":"modules/_favware_tslint_config.html#config.rules.object_literal_key_quotes","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":355,"kind":32,"name":"object-literal-shorthand","url":"modules/_favware_tslint_config.html#config.rules.object_literal_shorthand","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":356,"kind":32,"name":"object-literal-sort-keys","url":"modules/_favware_tslint_config.html#config.rules.object_literal_sort_keys","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":357,"kind":32,"name":"one-line","url":"modules/_favware_tslint_config.html#config.rules.one_line","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":358,"kind":32,"name":"only-arrow-functions","url":"modules/_favware_tslint_config.html#config.rules.only_arrow_functions","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":359,"kind":32,"name":"ordered-imports","url":"modules/_favware_tslint_config.html#config.rules.ordered_imports","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":360,"kind":32,"name":"prefer-conditional-expression","url":"modules/_favware_tslint_config.html#config.rules.prefer_conditional_expression","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":361,"kind":32,"name":"prefer-const","url":"modules/_favware_tslint_config.html#config.rules.prefer_const","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":362,"kind":32,"name":"prefer-for-of","url":"modules/_favware_tslint_config.html#config.rules.prefer_for_of","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":363,"kind":32,"name":"prefer-method-signature","url":"modules/_favware_tslint_config.html#config.rules.prefer_method_signature","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":364,"kind":32,"name":"prefer-object-spread","url":"modules/_favware_tslint_config.html#config.rules.prefer_object_spread","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":365,"kind":32,"name":"prefer-template","url":"modules/_favware_tslint_config.html#config.rules.prefer_template","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":366,"kind":32,"name":"promise-function-async","url":"modules/_favware_tslint_config.html#config.rules.promise_function_async","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":367,"kind":32,"name":"quotemark","url":"modules/_favware_tslint_config.html#config.rules.quotemark","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":368,"kind":32,"name":"radix","url":"modules/_favware_tslint_config.html#config.rules.radix","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":369,"kind":32,"name":"semicolon","url":"modules/_favware_tslint_config.html#config.rules.semicolon","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":370,"kind":32,"name":"space-before-function-paren","url":"modules/_favware_tslint_config.html#config.rules.space_before_function_paren","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":371,"kind":32,"name":"space-within-parens","url":"modules/_favware_tslint_config.html#config.rules.space_within_parens","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":372,"kind":32,"name":"switch-default","url":"modules/_favware_tslint_config.html#config.rules.switch_default","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":373,"kind":32,"name":"trailing-comma","url":"modules/_favware_tslint_config.html#config.rules.trailing_comma","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":374,"kind":32,"name":"triple-equals","url":"modules/_favware_tslint_config.html#config.rules.triple_equals","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":375,"kind":32,"name":"interface-over-type-literal","url":"modules/_favware_tslint_config.html#config.rules.interface_over_type_literal","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":376,"kind":32,"name":"unnecessary-else","url":"modules/_favware_tslint_config.html#config.rules.unnecessary_else","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":377,"kind":32,"name":"no-tautology-expression","url":"modules/_favware_tslint_config.html#config.rules.no_tautology_expression","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":378,"kind":32,"name":"static-this","url":"modules/_favware_tslint_config.html#config.rules.static_this","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":379,"kind":32,"name":"variable-name","url":"modules/_favware_tslint_config.html#config.rules.variable_name","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/tslint-config.config.rules"},{"id":380,"kind":2097152,"name":"charSets","url":"modules/_favware_unescape.html#charsets","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/unescape"},{"id":381,"kind":32,"name":"&quot;","url":"modules/_favware_unescape.html#charsets._quot_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":382,"kind":32,"name":"&#34;","url":"modules/_favware_unescape.html#charsets.__34_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":383,"kind":32,"name":"&#034;","url":"modules/_favware_unescape.html#charsets.__034_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":384,"kind":32,"name":"&#x22;","url":"modules/_favware_unescape.html#charsets.__x22_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":385,"kind":32,"name":"&apos;","url":"modules/_favware_unescape.html#charsets._apos_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":386,"kind":32,"name":"&#39;","url":"modules/_favware_unescape.html#charsets.__39_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":387,"kind":32,"name":"&#039;","url":"modules/_favware_unescape.html#charsets.__039_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":388,"kind":32,"name":"&#x27;","url":"modules/_favware_unescape.html#charsets.__x27_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":389,"kind":32,"name":"&amp;","url":"modules/_favware_unescape.html#charsets._amp_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":390,"kind":32,"name":"&#38;","url":"modules/_favware_unescape.html#charsets.__38_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":391,"kind":32,"name":"&#038;","url":"modules/_favware_unescape.html#charsets.__038_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":392,"kind":32,"name":"&#x26;","url":"modules/_favware_unescape.html#charsets.__x26_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":393,"kind":32,"name":"&gt;","url":"modules/_favware_unescape.html#charsets._gt_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":394,"kind":32,"name":"&#62;","url":"modules/_favware_unescape.html#charsets.__62_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":395,"kind":32,"name":"&#062;","url":"modules/_favware_unescape.html#charsets.__062_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":396,"kind":32,"name":"&#x3e;","url":"modules/_favware_unescape.html#charsets.__x3e_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":397,"kind":32,"name":"&lt;","url":"modules/_favware_unescape.html#charsets._lt_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":398,"kind":32,"name":"&#60;","url":"modules/_favware_unescape.html#charsets.__60_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":399,"kind":32,"name":"&#060;","url":"modules/_favware_unescape.html#charsets.__060_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":400,"kind":32,"name":"&#x3c;","url":"modules/_favware_unescape.html#charsets.__x3c_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":401,"kind":32,"name":"&copy;","url":"modules/_favware_unescape.html#charsets._copy_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":402,"kind":32,"name":"&#169;","url":"modules/_favware_unescape.html#charsets.__169_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":403,"kind":32,"name":"&#xa9;","url":"modules/_favware_unescape.html#charsets.__xa9_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":404,"kind":32,"name":"&reg;","url":"modules/_favware_unescape.html#charsets._reg_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":405,"kind":32,"name":"&#174;","url":"modules/_favware_unescape.html#charsets.__174_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":406,"kind":32,"name":"&#xae;","url":"modules/_favware_unescape.html#charsets.__xae_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":407,"kind":32,"name":"&cent;","url":"modules/_favware_unescape.html#charsets._cent_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":408,"kind":32,"name":"&#162;","url":"modules/_favware_unescape.html#charsets.__162_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":409,"kind":32,"name":"&#xa2;","url":"modules/_favware_unescape.html#charsets.__xa2_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":410,"kind":32,"name":"&euro;","url":"modules/_favware_unescape.html#charsets._euro_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":411,"kind":32,"name":"&#8364;","url":"modules/_favware_unescape.html#charsets.__8364_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":412,"kind":32,"name":"&pound;","url":"modules/_favware_unescape.html#charsets._pound_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":413,"kind":32,"name":"&#163;","url":"modules/_favware_unescape.html#charsets.__163_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":414,"kind":32,"name":"&#xa3;","url":"modules/_favware_unescape.html#charsets.__xa3_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":415,"kind":32,"name":"&yen;","url":"modules/_favware_unescape.html#charsets._yen_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":416,"kind":32,"name":"&#165;","url":"modules/_favware_unescape.html#charsets.__165_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":417,"kind":32,"name":"&#xa5;","url":"modules/_favware_unescape.html#charsets.__xa5_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":418,"kind":32,"name":"&nbsp;","url":"modules/_favware_unescape.html#charsets._nbsp_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":419,"kind":32,"name":"&#160;","url":"modules/_favware_unescape.html#charsets.__160_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":420,"kind":32,"name":"&#xa0;","url":"modules/_favware_unescape.html#charsets.__xa0_","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/unescape.charSets"},{"id":421,"kind":64,"name":"unescape","url":"modules/_favware_unescape.html#unescape","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/unescape"},{"id":422,"kind":256,"name":"FsOptions","url":"interfaces/_favware_yamlreader.fsoptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":423,"kind":1024,"name":"encoding","url":"interfaces/_favware_yamlreader.fsoptions.html#encoding","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/yamlreader.FsOptions"},{"id":424,"kind":1024,"name":"flag","url":"interfaces/_favware_yamlreader.fsoptions.html#flag","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/yamlreader.FsOptions"},{"id":425,"kind":128,"name":"YamlReaderError","url":"classes/_favware_yamlreader.yamlreadererror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":426,"kind":512,"name":"constructor","url":"classes/_favware_yamlreader.yamlreadererror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"@favware/yamlreader.YamlReaderError"},{"id":427,"kind":1024,"name":"name","url":"classes/_favware_yamlreader.yamlreadererror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/yamlreader.YamlReaderError"},{"id":428,"kind":1024,"name":"message","url":"classes/_favware_yamlreader.yamlreadererror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/yamlreader.YamlReaderError"},{"id":429,"kind":1024,"name":"stack","url":"classes/_favware_yamlreader.yamlreadererror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/yamlreader.YamlReaderError"},{"id":430,"kind":1024,"name":"Error","url":"classes/_favware_yamlreader.yamlreadererror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/yamlreader.YamlReaderError"},{"id":431,"kind":4194304,"name":"YamlReaderOptions","url":"modules/_favware_yamlreader.html#yamlreaderoptions","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":432,"kind":64,"name":"readYaml","url":"modules/_favware_yamlreader.html#readyaml","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/yamlreader"},{"id":433,"kind":4194304,"name":"PrismVideoFormat","url":"modules/_favware_ytdl_prismplayer.html#prismvideoformat","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"@favware/ytdl-prismplayer"},{"id":434,"kind":64,"name":"play","url":"modules/_favware_ytdl_prismplayer.html#play","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/ytdl-prismplayer"},{"id":435,"kind":256,"name":"Chars","url":"interfaces/_favware_zalgo.chars.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":436,"kind":1024,"name":"all","url":"interfaces/_favware_zalgo.chars.html#all","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":437,"kind":1024,"name":"up","url":"interfaces/_favware_zalgo.chars.html#up","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":438,"kind":1024,"name":"middle","url":"interfaces/_favware_zalgo.chars.html#middle","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":439,"kind":1024,"name":"down","url":"interfaces/_favware_zalgo.chars.html#down","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":440,"kind":1024,"name":"pattern","url":"interfaces/_favware_zalgo.chars.html#pattern","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.Chars"},{"id":441,"kind":256,"name":"ZalgoOptions","url":"interfaces/_favware_zalgo.zalgooptions.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":442,"kind":1024,"name":"up","url":"interfaces/_favware_zalgo.zalgooptions.html#up","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":443,"kind":1024,"name":"middle","url":"interfaces/_favware_zalgo.zalgooptions.html#middle","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":444,"kind":1024,"name":"down","url":"interfaces/_favware_zalgo.zalgooptions.html#down","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":445,"kind":1024,"name":"size","url":"interfaces/_favware_zalgo.zalgooptions.html#size","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoOptions"},{"id":446,"kind":256,"name":"ZalgoCountsMap","url":"interfaces/_favware_zalgo.zalgocountsmap.html","classes":"tsd-kind-interface tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":447,"kind":1024,"name":"up","url":"interfaces/_favware_zalgo.zalgocountsmap.html#up","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoCountsMap"},{"id":448,"kind":1024,"name":"middle","url":"interfaces/_favware_zalgo.zalgocountsmap.html#middle","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoCountsMap"},{"id":449,"kind":1024,"name":"down","url":"interfaces/_favware_zalgo.zalgocountsmap.html#down","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"@favware/zalgo.ZalgoCountsMap"},{"id":450,"kind":2097152,"name":"chars","url":"modules/_favware_zalgo.html#chars-1","classes":"tsd-kind-object-literal tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":451,"kind":32,"name":"up","url":"modules/_favware_zalgo.html#chars-1.up","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/zalgo.chars"},{"id":452,"kind":32,"name":"middle","url":"modules/_favware_zalgo.html#chars-1.middle","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/zalgo.chars"},{"id":453,"kind":32,"name":"down","url":"modules/_favware_zalgo.html#chars-1.down","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"@favware/zalgo.chars"},{"id":454,"kind":128,"name":"ZalgoError","url":"classes/_favware_zalgo.zalgoerror.html","classes":"tsd-kind-class tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":455,"kind":512,"name":"constructor","url":"classes/_favware_zalgo.zalgoerror.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"@favware/zalgo.ZalgoError"},{"id":456,"kind":1024,"name":"name","url":"classes/_favware_zalgo.zalgoerror.html#name","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/zalgo.ZalgoError"},{"id":457,"kind":1024,"name":"message","url":"classes/_favware_zalgo.zalgoerror.html#message","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-inherited","parent":"@favware/zalgo.ZalgoError"},{"id":458,"kind":1024,"name":"stack","url":"classes/_favware_zalgo.zalgoerror.html#stack","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-overwrite tsd-is-inherited","parent":"@favware/zalgo.ZalgoError"},{"id":459,"kind":1024,"name":"Error","url":"classes/_favware_zalgo.zalgoerror.html#error","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"@favware/zalgo.ZalgoError"},{"id":460,"kind":64,"name":"unicodeStringSplitter","url":"modules/_favware_zalgo.html#unicodestringsplitter","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":461,"kind":64,"name":"hasOwnProperty","url":"modules/_favware_zalgo.html#hasownproperty","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"@favware/zalgo"},{"id":462,"kind":64,"name":"randomizer","url":"modules/_favware_zalgo.html#randomizer","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":463,"kind":64,"name":"zalgo","url":"modules/_favware_zalgo.html#zalgo","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":464,"kind":64,"name":"banish","url":"modules/_favware_zalgo.html#banish","classes":"tsd-kind-function tsd-parent-kind-module","parent":"@favware/zalgo"},{"id":465,"kind":2,"name":"@favware/converter","url":"modules/_favware_converter.html","classes":"tsd-kind-module"},{"id":466,"kind":2,"name":"@favware/crypto","url":"modules/_favware_crypto.html","classes":"tsd-kind-module"},{"id":467,"kind":2,"name":"@favware/eslint-config","url":"modules/_favware_eslint_config.html","classes":"tsd-kind-module"},{"id":468,"kind":2,"name":"@favware/milky-tslint","url":"modules/_favware_milky_tslint.html","classes":"tsd-kind-module"},{"id":469,"kind":2,"name":"@favware/querystring","url":"modules/_favware_querystring.html","classes":"tsd-kind-module"},{"id":470,"kind":2,"name":"@favware/tslint-config","url":"modules/_favware_tslint_config.html","classes":"tsd-kind-module"},{"id":471,"kind":2,"name":"@favware/unescape","url":"modules/_favware_unescape.html","classes":"tsd-kind-module"},{"id":472,"kind":2,"name":"@favware/yamlreader","url":"modules/_favware_yamlreader.html","classes":"tsd-kind-module"},{"id":473,"kind":2,"name":"@favware/ytdl-prismplayer","url":"modules/_favware_ytdl_prismplayer.html","classes":"tsd-kind-module"},{"id":474,"kind":2,"name":"@favware/zalgo","url":"modules/_favware_zalgo.html","classes":"tsd-kind-module"}]};
