+++ docs/enums/_favware_converter.system.html
@@ -70,17 +70,6 @@
 					<div class="lead">
 						<p>Predefined options for the unit system</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{System} System unit system enum</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{System.PRIMARY} Primary for when the unit is of the primary system</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{System.SECONDARY} Secondary for when the unit is of the secondary system</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel-group tsd-index-group">
@@ -105,9 +94,14 @@
 					<div class="tsd-signature tsd-kind-icon">PRIMARY<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;primary&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L57">packages/converter/src/interfaces.ts:57</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L45">packages/converter/src/interfaces.ts:45</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>For when the unit is of the primary system</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-enum-member tsd-parent-kind-enum">
 					<a name="secondary" class="tsd-anchor"></a>
@@ -115,9 +109,14 @@
 					<div class="tsd-signature tsd-kind-icon">SECONDARY<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;secondary&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L58">packages/converter/src/interfaces.ts:58</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L47">packages/converter/src/interfaces.ts:47</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>For when the unit is of the secondary system</p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -134,10 +133,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
