+++ docs/enums/_favware_crypto.radix.html
@@ -70,23 +70,6 @@
 					<div class="lead">
 						<p>Possible options to use as the radix</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>enum</dt>
-						<dd><p>{Radix} Radix</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{Radix.base16} 16</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{Radix.base32} 32</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{Radix.base58} 58</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{Radix.base64} 64</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel-group tsd-index-group">
@@ -114,7 +97,7 @@
 					<div class="tsd-signature tsd-kind-icon">base16<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;16</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L9">packages/crypto/src/util.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L2">packages/crypto/src/util.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -124,7 +107,7 @@
 					<div class="tsd-signature tsd-kind-icon">base32<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;32</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L9">packages/crypto/src/util.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L2">packages/crypto/src/util.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -134,7 +117,7 @@
 					<div class="tsd-signature tsd-kind-icon">base36<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;36</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L9">packages/crypto/src/util.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L2">packages/crypto/src/util.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -144,7 +127,7 @@
 					<div class="tsd-signature tsd-kind-icon">base58<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;58</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L9">packages/crypto/src/util.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L2">packages/crypto/src/util.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -154,7 +137,7 @@
 					<div class="tsd-signature tsd-kind-icon">base64<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;64</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L9">packages/crypto/src/util.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L2">packages/crypto/src/util.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -173,10 +156,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
