+++ docs/enums/_favware_querystring.functionsource.html
@@ -70,17 +70,6 @@
 					<div class="lead">
 						<p>Predefined potential sources for errors</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{FunctionSource} FunctionSource Source Reference Enum</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{FunctionSource.Parse} Parse for when the source is <a href="../modules/_favware_querystring.html#parse">parse</a></p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{FunctionSource.Stringify} Stringify for when the source is <a href="../modules/_favware_querystring.html#stringify">stringify</a></p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel-group tsd-index-group">
@@ -105,9 +94,14 @@
 					<div class="tsd-signature tsd-kind-icon">Parse<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;parse&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/querystring/src/types.ts#L53">packages/querystring/src/types.ts:53</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/querystring/src/types.ts#L34">packages/querystring/src/types.ts:34</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>For when the source is <a href="../modules/_favware_querystring.html#parse">parse</a></p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-enum-member tsd-parent-kind-enum">
 					<a name="stringify" class="tsd-anchor"></a>
@@ -115,9 +109,14 @@
 					<div class="tsd-signature tsd-kind-icon">Stringify<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;stringify&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/querystring/src/types.ts#L54">packages/querystring/src/types.ts:54</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/querystring/src/types.ts#L36">packages/querystring/src/types.ts:36</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>For when the source is <a href="../modules/_favware_querystring.html#stringify">stringify</a></p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -134,10 +133,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class="current tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
@@ -157,7 +162,7 @@
 				<ul class="before-current">
 				</ul>
 				<ul class="current">
-					<li class="current tsd-kind-enum tsd-parent-kind-module tsd-is-private">
+					<li class="current tsd-kind-enum tsd-parent-kind-module">
 						<a href="_favware_querystring.functionsource.html" class="tsd-kind-icon">Function<wbr>Source</a>
 						<ul>
 							<li class=" tsd-kind-enum-member tsd-parent-kind-enum">
@@ -173,13 +178,13 @@
 					<li class=" tsd-kind-interface tsd-parent-kind-module">
 						<a href="../interfaces/_favware_querystring.parseoptions.html" class="tsd-kind-icon">Parse<wbr>Options</a>
 					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-module tsd-is-private">
+					<li class=" tsd-kind-interface tsd-parent-kind-module">
 						<a href="../interfaces/_favware_querystring.querystringobject.html" class="tsd-kind-icon">Querystring<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-module">
 						<a href="../interfaces/_favware_querystring.stringifyoptions.html" class="tsd-kind-icon">Stringify<wbr>Options</a>
 					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-is-private">
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module">
 						<a href="../modules/_favware_querystring.html#querystringerror" class="tsd-kind-icon">Querystring<wbr>Error</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-module">
