+++ docs/globals.html
@@ -71,6 +71,7 @@
 								<li class="tsd-kind-module"><a href="modules/_favware_eslint_config.html" class="tsd-kind-icon">@favware/eslint-<wbr>config</a></li>
 								<li class="tsd-kind-module"><a href="modules/_favware_milky_tslint.html" class="tsd-kind-icon">@favware/milky-<wbr>tslint</a></li>
 								<li class="tsd-kind-module"><a href="modules/_favware_querystring.html" class="tsd-kind-icon">@favware/querystring</a></li>
+								<li class="tsd-kind-module"><a href="modules/_favware_tslint_config.html" class="tsd-kind-icon">@favware/tslint-<wbr>config</a></li>
 								<li class="tsd-kind-module"><a href="modules/_favware_unescape.html" class="tsd-kind-icon">@favware/unescape</a></li>
 								<li class="tsd-kind-module"><a href="modules/_favware_yamlreader.html" class="tsd-kind-icon">@favware/yamlreader</a></li>
 								<li class="tsd-kind-module"><a href="modules/_favware_ytdl_prismplayer.html" class="tsd-kind-icon">@favware/ytdl-<wbr>prismplayer</a></li>
@@ -101,6 +102,9 @@
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="modules/_favware_unescape.html">@favware/unescape</a>
