+++ docs/interfaces/_favware_converter.measurementsystemdata.html
@@ -70,20 +70,6 @@
 					<div class="lead">
 						<p>Measure system data</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{MeasurementSystemData} IMeasurementSystemData measure system data for primary and secondary</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{string} default The default unit to use</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{number} ratio The ratio to convert to other measurement system</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{function} [transform] A transformation function when a simple ratio is not enough</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel tsd-hierarchy">
@@ -122,9 +108,14 @@
 					<div class="tsd-signature tsd-kind-icon">default<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L9">packages/converter/src/interfaces.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L4">packages/converter/src/interfaces.ts:4</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>The default unit to use</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="ratio" class="tsd-anchor"></a>
@@ -132,9 +123,14 @@
 					<div class="tsd-signature tsd-kind-icon">ratio<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L10">packages/converter/src/interfaces.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L6">packages/converter/src/interfaces.ts:6</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>The ratio to convert to other measurement system</p>
+						</div>
+					</div>
 				</section>
 			</section>
 			<section class="tsd-panel-group tsd-member-group ">
@@ -149,9 +145,14 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L12">packages/converter/src/interfaces.ts:12</a></li>
+									<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L9">packages/converter/src/interfaces.ts:9</a></li>
 								</ul>
 							</aside>
+							<div class="tsd-comment tsd-typography">
+								<div class="lead">
+									<p>A transformation function when a simple ratio is not enough</p>
+								</div>
+							</div>
 							<h4 class="tsd-parameters-title">Parameters</h4>
 							<ul class="tsd-parameters">
 								<li>
@@ -177,10 +178,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
