+++ docs/interfaces/_favware_converter.unitdata.html
@@ -70,23 +70,6 @@
 					<div class="lead">
 						<p>Single property of definitions</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{UnitData} IUnitData unit definition property</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{string} id ID of the unit</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{number} multiplier Multiplier for the unit</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{number} valueShift A zero value shift to use instead of a multiplier</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{boolean} isSecondary whether this data prop belongs to a secondary system</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel tsd-hierarchy">
@@ -121,9 +104,14 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L35">packages/converter/src/interfaces.ts:35</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L27">packages/converter/src/interfaces.ts:27</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>ID of the unit</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="multiplier" class="tsd-anchor"></a>
@@ -131,9 +119,14 @@
 					<div class="tsd-signature tsd-kind-icon">multiplier<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L37">packages/converter/src/interfaces.ts:37</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L31">packages/converter/src/interfaces.ts:31</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Multiplier for the unit</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="system" class="tsd-anchor"></a>
@@ -141,9 +134,14 @@
 					<div class="tsd-signature tsd-kind-icon">system<span class="tsd-signature-symbol">:</span> <a href="../enums/_favware_converter.system.html" class="tsd-signature-type">System</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L36">packages/converter/src/interfaces.ts:36</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L29">packages/converter/src/interfaces.ts:29</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether the unit is of primary or secondary system</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="valueshift" class="tsd-anchor"></a>
@@ -151,9 +149,14 @@
 					<div class="tsd-signature tsd-kind-icon">value<wbr>Shift<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L38">packages/converter/src/interfaces.ts:38</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L33">packages/converter/src/interfaces.ts:33</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>A zero value shift to use instead of a multiplier</p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -170,10 +173,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
