+++ docs/interfaces/_favware_converter.unitdefinition.html
@@ -70,11 +70,6 @@
 					<div class="lead">
 						<p>Array of supported unit definitions</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{UnitDefinition} IUnitDefinition array of all definitions</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel tsd-hierarchy">
@@ -109,9 +104,14 @@
 					<div class="tsd-signature tsd-kind-icon">data<span class="tsd-signature-symbol">:</span> <a href="_favware_converter.unitdata.html" class="tsd-signature-type">UnitData</a><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L23">packages/converter/src/interfaces.ts:23</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L21">packages/converter/src/interfaces.ts:21</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Unit data that belongs to this unit&#39;s definition</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="name" class="tsd-anchor"></a>
@@ -119,9 +119,14 @@
 					<div class="tsd-signature tsd-kind-icon">name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L20">packages/converter/src/interfaces.ts:20</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L15">packages/converter/src/interfaces.ts:15</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Name of the unit definition</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="primary" class="tsd-anchor"></a>
@@ -129,9 +134,14 @@
 					<div class="tsd-signature tsd-kind-icon">primary<span class="tsd-signature-symbol">:</span> <a href="_favware_converter.measurementsystemdata.html" class="tsd-signature-type">MeasurementSystemData</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L21">packages/converter/src/interfaces.ts:21</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L17">packages/converter/src/interfaces.ts:17</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Primary measurement system data</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="secondary" class="tsd-anchor"></a>
@@ -139,9 +149,14 @@
 					<div class="tsd-signature tsd-kind-icon">secondary<span class="tsd-signature-symbol">:</span> <a href="_favware_converter.measurementsystemdata.html" class="tsd-signature-type">MeasurementSystemData</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/converter/src/interfaces.ts#L22">packages/converter/src/interfaces.ts:22</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/converter/src/interfaces.ts#L19">packages/converter/src/interfaces.ts:19</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Secondary measurement system data</p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -158,10 +173,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
