+++ docs/interfaces/_favware_milky_tslint.reportoptions.html
@@ -97,9 +97,14 @@
 					<div class="tsd-signature tsd-kind-icon">allow<wbr>Warnings<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/types.ts#L17">packages/milky-tslint/src/types.ts:17</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/types.ts#L29">packages/milky-tslint/src/types.ts:29</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether to allow lint warnings</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="emiterror" class="tsd-anchor"></a>
@@ -107,9 +112,14 @@
 					<div class="tsd-signature tsd-kind-icon">emit<wbr>Error<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/types.ts#L14">packages/milky-tslint/src/types.ts:14</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/types.ts#L23">packages/milky-tslint/src/types.ts:23</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether to emit errors</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="reportlimit" class="tsd-anchor"></a>
@@ -117,9 +127,14 @@
 					<div class="tsd-signature tsd-kind-icon">report<wbr>Limit<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/types.ts#L15">packages/milky-tslint/src/types.ts:15</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/types.ts#L25">packages/milky-tslint/src/types.ts:25</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>The limit of the amount of reports to return</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="summarizefailureoutput" class="tsd-anchor"></a>
@@ -127,9 +142,14 @@
 					<div class="tsd-signature tsd-kind-icon">summarize<wbr>Failure<wbr>Output<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/types.ts#L16">packages/milky-tslint/src/types.ts:16</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/types.ts#L27">packages/milky-tslint/src/types.ts:27</a></li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether to summarize the failure output after run</p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -145,11 +165,17 @@
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
 					<li class="current tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
@@ -205,6 +231,9 @@
 					</li>
 					<li class=" tsd-kind-variable tsd-parent-kind-module">
 						<a href="../modules/_favware_milky_tslint.html#eventemitter" class="tsd-kind-icon">eventemitter</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module">
+						<a href="../modules/_favware_milky_tslint.html#isstring" class="tsd-kind-icon">is<wbr>String</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-module">
 						<a href="../modules/_favware_milky_tslint.html#milkylint" class="tsd-kind-icon">milky<wbr>Lint</a>
