+++ docs/interfaces/_favware_querystring.querystringobject.html
@@ -112,6 +112,9 @@
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
 					</li>
 					<li class=" tsd-kind-module">
