+++ docs/interfaces/_favware_zalgo.chars.html
@@ -70,26 +70,6 @@
 					<div class="lead">
 						<p>Interface for the characters object</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{Chars} IChars characters object</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{string[]} [all] All characters</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{string[]} up characters going up</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{string[]} middle characters in the middle</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{string[]} down characters going below</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{RegExp} [pattern] Stored RegExp pattern</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel tsd-hierarchy">
@@ -105,7 +85,7 @@
 				<div class="tsd-signature tsd-kind-icon"><span class="tsd-signature-symbol">[</span>indexSignature: <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">]: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">RegExp</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">undefined</span></div>
 				<div class="tsd-comment tsd-typography">
 					<div class="lead">
-						<p>Interface for the characters object</p>
+						<p>Index Signature for TS parsing</p>
 					</div>
 				</div>
 			</section>
@@ -134,9 +114,14 @@
 					<div class="tsd-signature tsd-kind-icon">all<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/chars.ts#L11">packages/zalgo/src/chars.ts:11</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:4</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>All characters</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="down" class="tsd-anchor"></a>
@@ -144,9 +129,14 @@
 					<div class="tsd-signature tsd-kind-icon">down<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/chars.ts#L14">packages/zalgo/src/chars.ts:14</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:10</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Characters going below</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="middle" class="tsd-anchor"></a>
@@ -154,9 +144,14 @@
 					<div class="tsd-signature tsd-kind-icon">middle<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/chars.ts#L13">packages/zalgo/src/chars.ts:13</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:8</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Characters in the middle</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="pattern" class="tsd-anchor"></a>
@@ -164,9 +159,14 @@
 					<div class="tsd-signature tsd-kind-icon">pattern<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">RegExp</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/chars.ts#L15">packages/zalgo/src/chars.ts:15</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:12</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Stored RegExp pattern</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="up" class="tsd-anchor"></a>
@@ -174,9 +174,14 @@
 					<div class="tsd-signature tsd-kind-icon">up<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/chars.ts#L12">packages/zalgo/src/chars.ts:12</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:6</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Characters going up</p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -193,10 +198,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
@@ -214,6 +225,9 @@
 			</nav>
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
+					<li class=" tsd-kind-class tsd-parent-kind-module">
+						<a href="../classes/_favware_zalgo.zalgoerror.html" class="tsd-kind-icon">Zalgo<wbr>Error</a>
+					</li>
 				</ul>
 				<ul class="current">
 					<li class="current tsd-kind-interface tsd-parent-kind-module">
@@ -239,10 +253,22 @@
 				</ul>
 				<ul class="after-current">
 					<li class=" tsd-kind-interface tsd-parent-kind-module">
+						<a href="_favware_zalgo.zalgocountsmap.html" class="tsd-kind-icon">Zalgo<wbr>Counts<wbr>Map</a>
+					</li>
+					<li class=" tsd-kind-interface tsd-parent-kind-module">
 						<a href="_favware_zalgo.zalgooptions.html" class="tsd-kind-icon">Zalgo<wbr>Options</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-module">
 						<a href="../modules/_favware_zalgo.html#banish" class="tsd-kind-icon">banish</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/_favware_zalgo.html#hasownproperty" class="tsd-kind-icon">has<wbr>Own<wbr>Property</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module">
+						<a href="../modules/_favware_zalgo.html#randomizer" class="tsd-kind-icon">randomizer</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module">
+						<a href="../modules/_favware_zalgo.html#unicodestringsplitter" class="tsd-kind-icon">unicode<wbr>String<wbr>Splitter</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-module">
 						<a href="../modules/_favware_zalgo.html#zalgo" class="tsd-kind-icon">zalgo</a>
