+++ docs/interfaces/_favware_zalgo.zalgooptions.html
@@ -70,23 +70,6 @@
 					<div class="lead">
 						<p>Options for the Zalgo screw up</p>
 					</div>
-					<dl class="tsd-comment-tags">
-						<dt>typedef</dt>
-						<dd><p>{ZalgoOptions} IZalgoOptions Options for the Zalgo</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{boolean} up Whether the zalgo should go up</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{boolean} middle Whether the zalgo should go in the middle</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{boolean} down Whether the zalgo should go downards</p>
-						</dd>
-						<dt>property</dt>
-						<dd><p>{&#39;mini&#39; | &#39;maxi&#39;} size Whether the zalgo should be mini or maxi</p>
-						</dd>
-					</dl>
 				</div>
 			</section>
 			<section class="tsd-panel tsd-hierarchy">
@@ -121,9 +104,14 @@
 					<div class="tsd-signature tsd-kind-icon">down<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/zalgo.ts#L14">packages/zalgo/src/zalgo.ts:14</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:25</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether the zalgo should go downards</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="middle" class="tsd-anchor"></a>
@@ -131,9 +119,14 @@
 					<div class="tsd-signature tsd-kind-icon">middle<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/zalgo.ts#L13">packages/zalgo/src/zalgo.ts:13</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:23</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether the zalgo should go in the middle</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="size" class="tsd-anchor"></a>
@@ -141,9 +134,14 @@
 					<div class="tsd-signature tsd-kind-icon">size<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">"mini"</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">"maxi"</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">&quot;&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/zalgo.ts#L15">packages/zalgo/src/zalgo.ts:15</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:27</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether the zalgo should be mini or maxi</p>
+						</div>
+					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="up" class="tsd-anchor"></a>
@@ -151,9 +149,14 @@
 					<div class="tsd-signature tsd-kind-icon">up<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/zalgo/src/zalgo.ts#L12">packages/zalgo/src/zalgo.ts:12</a></li>
+							<li>Defined in packages/zalgo/src/interfaces.ts:21</li>
 						</ul>
 					</aside>
+					<div class="tsd-comment tsd-typography">
+						<div class="lead">
+							<p>Whether the zalgo should go up</p>
+						</div>
+					</div>
 				</section>
 			</section>
 		</div>
@@ -170,10 +173,16 @@
 						<a href="../modules/_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="../modules/_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="../modules/_favware_unescape.html">@favware/unescape</a>
@@ -191,8 +200,14 @@
 			</nav>
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
+					<li class=" tsd-kind-class tsd-parent-kind-module">
+						<a href="../classes/_favware_zalgo.zalgoerror.html" class="tsd-kind-icon">Zalgo<wbr>Error</a>
+					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-module">
 						<a href="_favware_zalgo.chars.html" class="tsd-kind-icon">Chars</a>
+					</li>
+					<li class=" tsd-kind-interface tsd-parent-kind-module">
+						<a href="_favware_zalgo.zalgocountsmap.html" class="tsd-kind-icon">Zalgo<wbr>Counts<wbr>Map</a>
 					</li>
 				</ul>
 				<ul class="current">
@@ -217,6 +232,15 @@
 				<ul class="after-current">
 					<li class=" tsd-kind-function tsd-parent-kind-module">
 						<a href="../modules/_favware_zalgo.html#banish" class="tsd-kind-icon">banish</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/_favware_zalgo.html#hasownproperty" class="tsd-kind-icon">has<wbr>Own<wbr>Property</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module">
+						<a href="../modules/_favware_zalgo.html#randomizer" class="tsd-kind-icon">randomizer</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module">
+						<a href="../modules/_favware_zalgo.html#unicodestringsplitter" class="tsd-kind-icon">unicode<wbr>String<wbr>Splitter</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-module">
 						<a href="../modules/_favware_zalgo.html#zalgo" class="tsd-kind-icon">zalgo</a>
