+++ docs/modules/_favware_crypto.html
@@ -164,7 +164,7 @@ crypto(<span class="hljs-number">160</span>, Radix.base64);</code></pre>
 					<div class="tsd-signature tsd-kind-icon">CHARACTERS<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> =&nbsp;&#x27;ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/&#x27;.split(&#x27;&#x27;)</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L36">packages/crypto/src/util.ts:36</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L15">packages/crypto/src/util.ts:15</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -174,23 +174,9 @@ crypto(<span class="hljs-number">160</span>, Radix.base64);</code></pre>
 					<div class="tsd-signature tsd-kind-icon">DEFAULT_<wbr>BITS<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">36</span><span class="tsd-signature-symbol"> =&nbsp;36</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L26">packages/crypto/src/util.ts:26</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L13">packages/crypto/src/util.ts:13</a></li>
 						</ul>
 					</aside>
-					<div class="tsd-comment tsd-typography">
-						<dl class="tsd-comment-tags">
-							<dt>constant</dt>
-							<dd></dd>
-							<dt>type</dt>
-							<dd><p>{number}</p>
-							</dd>
-							<dt>default</dt>
-							<dd></dd>
-							<dt>name</dt>
-							<dd><p>DEFAULT_BITS</p>
-							</dd>
-						</dl>
-					</div>
 				</section>
 				<section class="tsd-panel tsd-member tsd-kind-variable tsd-parent-kind-module">
 					<a name="default_radix" class="tsd-anchor"></a>
@@ -198,23 +184,9 @@ crypto(<span class="hljs-number">160</span>, Radix.base64);</code></pre>
 					<div class="tsd-signature tsd-kind-icon">DEFAULT_<wbr>RADIX<span class="tsd-signature-symbol">:</span> <a href="../enums/_favware_crypto.radix.html#base64" class="tsd-signature-type">base64</a><span class="tsd-signature-symbol"> =&nbsp;Radix.base64</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/util.ts#L34">packages/crypto/src/util.ts:34</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/util.ts#L14">packages/crypto/src/util.ts:14</a></li>
 						</ul>
 					</aside>
-					<div class="tsd-comment tsd-typography">
-						<dl class="tsd-comment-tags">
-							<dt>constant</dt>
-							<dd></dd>
-							<dt>type</dt>
-							<dd><p>{number}</p>
-							</dd>
-							<dt>default</dt>
-							<dd></dd>
-							<dt>name</dt>
-							<dd><p>DEFAULT_RADIX</p>
-							</dd>
-						</dl>
-					</div>
 				</section>
 			</section>
 			<section class="tsd-panel-group tsd-member-group ">
@@ -229,7 +201,7 @@ crypto(<span class="hljs-number">160</span>, Radix.base64);</code></pre>
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/crypto/src/crypto.ts#L13">packages/crypto/src/crypto.ts:13</a></li>
+									<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/crypto/src/crypto.ts#L12">packages/crypto/src/crypto.ts:12</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -241,9 +213,19 @@ crypto(<span class="hljs-number">160</span>, Radix.base64);</code></pre>
 							<ul class="tsd-parameters">
 								<li>
 									<h5><span class="tsd-flag ts-flagDefault value">Default value</span> bits: <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> =&nbsp;DEFAULT_BITS</span></h5>
+									<div class="tsd-comment tsd-typography">
+										<div class="lead">
+											<p>The amount of bits to generate</p>
+										</div>
+									</div>
 								</li>
 								<li>
 									<h5><span class="tsd-flag ts-flagDefault value">Default value</span> radix: <a href="../enums/_favware_crypto.radix.html" class="tsd-signature-type">Radix</a><span class="tsd-signature-symbol"> =&nbsp;DEFAULT_RADIX</span></h5>
+									<div class="tsd-comment tsd-typography">
+										<div class="lead">
+											<p>The radix to generate for</p>
+										</div>
+									</div>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -266,10 +248,16 @@ crypto(<span class="hljs-number">160</span>, Radix.base64);</code></pre>
 						<a href="_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="_favware_unescape.html">@favware/unescape</a>
