+++ docs/modules/_favware_milky_tslint.html
@@ -256,6 +256,7 @@ gulp.task(<span class="hljs-string">"lint"</span>, () =&gt; {
 						<section class="tsd-index-section ">
 							<h3>Functions</h3>
 							<ul class="tsd-index-list">
+								<li class="tsd-kind-function tsd-parent-kind-module"><a href="_favware_milky_tslint.html#isstring" class="tsd-kind-icon">is<wbr>String</a></li>
 								<li class="tsd-kind-function tsd-parent-kind-module"><a href="_favware_milky_tslint.html#milkylint" class="tsd-kind-icon">milky<wbr>Lint</a></li>
 								<li class="tsd-kind-function tsd-parent-kind-module"><a href="_favware_milky_tslint.html#milkyreport" class="tsd-kind-icon">milky<wbr>Report</a></li>
 							</ul>
@@ -271,13 +272,41 @@ gulp.task(<span class="hljs-string">"lint"</span>, () =&gt; {
 					<div class="tsd-signature tsd-kind-icon">eventemitter<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">StrictEventEmitter</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">EventEmitter</span><span class="tsd-signature-symbol">, </span><a href="../interfaces/_favware_milky_tslint.milkyevents.html" class="tsd-signature-type">MilkyEvents</a><span class="tsd-signature-symbol">&gt;</span><span class="tsd-signature-symbol"> =&nbsp;new EventEmitter()</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/events.ts#L24">packages/milky-tslint/src/events.ts:24</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/events.ts#L16">packages/milky-tslint/src/events.ts:16</a></li>
 						</ul>
 					</aside>
 				</section>
 			</section>
 			<section class="tsd-panel-group tsd-member-group ">
 				<h2>Functions</h2>
+				<section class="tsd-panel tsd-member tsd-kind-function tsd-parent-kind-module">
+					<a name="isstring" class="tsd-anchor"></a>
+					<h3><span class="tsd-flag ts-flagConst">Const</span> is<wbr>String</h3>
+					<ul class="tsd-signatures tsd-kind-function tsd-parent-kind-module">
+						<li class="tsd-signature tsd-kind-icon">is<wbr>String<span class="tsd-signature-symbol">(</span>str<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">unknown</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">boolean</span></li>
+					</ul>
+					<ul class="tsd-descriptions">
+						<li class="tsd-description">
+							<aside class="tsd-sources">
+								<ul>
+									<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/milky.ts#L15">packages/milky-tslint/src/milky.ts:15</a></li>
+								</ul>
+							</aside>
+							<div class="tsd-comment tsd-typography">
+								<div class="lead">
+									<p>TypeGuard to ensure param is string</p>
+								</div>
+							</div>
+							<h4 class="tsd-parameters-title">Parameters</h4>
+							<ul class="tsd-parameters">
+								<li>
+									<h5>str: <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">unknown</span></h5>
+								</li>
+							</ul>
+							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">boolean</span></h4>
+						</li>
+					</ul>
+				</section>
 				<section class="tsd-panel tsd-member tsd-kind-function tsd-parent-kind-module">
 					<a name="milkylint" class="tsd-anchor"></a>
 					<h3><span class="tsd-flag ts-flagConst">Const</span> milky<wbr>Lint</h3>
@@ -288,16 +317,21 @@ gulp.task(<span class="hljs-string">"lint"</span>, () =&gt; {
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/milky.ts#L66">packages/milky-tslint/src/milky.ts:66</a></li>
+									<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/milky.ts#L56">packages/milky-tslint/src/milky.ts:56</a></li>
 								</ul>
 							</aside>
+							<div class="tsd-comment tsd-typography">
+								<div class="lead">
+									<p>TSLint plugin for Gulp</p>
+								</div>
+							</div>
 							<h4 class="tsd-parameters-title">Parameters</h4>
 							<ul class="tsd-parameters">
 								<li>
 									<h5><span class="tsd-flag ts-flagDefault value">Default value</span> pluginOptions: <a href="../interfaces/_favware_milky_tslint.pluginoptions.html" class="tsd-signature-type">PluginOptions</a><span class="tsd-signature-symbol"> =&nbsp;{}</span></h5>
 									<div class="tsd-comment tsd-typography">
 										<div class="lead">
-											<p>Options to pass to the plugin</p>
+											<p>Options to pass to the plugin, see <a href="../interfaces/_favware_milky_tslint.pluginoptions.html">PluginOptions</a></p>
 										</div>
 									</div>
 								</li>
@@ -316,16 +350,21 @@ gulp.task(<span class="hljs-string">"lint"</span>, () =&gt; {
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/milky-tslint/src/milky.ts#L104">packages/milky-tslint/src/milky.ts:104</a></li>
+									<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/milky-tslint/src/milky.ts#L93">packages/milky-tslint/src/milky.ts:93</a></li>
 								</ul>
 							</aside>
+							<div class="tsd-comment tsd-typography">
+								<div class="lead">
+									<p>Reporter for milky-tslint</p>
+								</div>
+							</div>
 							<h4 class="tsd-parameters-title">Parameters</h4>
 							<ul class="tsd-parameters">
 								<li>
 									<h5><span class="tsd-flag ts-flagDefault value">Default value</span> options: <a href="../interfaces/_favware_milky_tslint.reportoptions.html" class="tsd-signature-type">ReportOptions</a><span class="tsd-signature-symbol"> =&nbsp;{emitError: false, reportLimit: 0, summarizeFailureOutput: true, allowWarnings: false,}</span></h5>
 									<div class="tsd-comment tsd-typography">
 										<div class="lead">
-											<p>Reporter options</p>
+											<p>options for the report, see <a href="../interfaces/_favware_milky_tslint.reportoptions.html">ReportOptions</a></p>
 										</div>
 									</div>
 								</li>
@@ -348,11 +387,17 @@ gulp.task(<span class="hljs-string">"lint"</span>, () =&gt; {
 					<li class=" tsd-kind-module">
 						<a href="_favware_crypto.html">@favware/crypto</a>
 					</li>
+					<li class=" tsd-kind-module">
+						<a href="_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
 					<li class="current tsd-kind-module">
 						<a href="_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="_favware_unescape.html">@favware/unescape</a>
@@ -390,6 +435,9 @@ gulp.task(<span class="hljs-string">"lint"</span>, () =&gt; {
 					</li>
 					<li class=" tsd-kind-variable tsd-parent-kind-module">
 						<a href="_favware_milky_tslint.html#eventemitter" class="tsd-kind-icon">eventemitter</a>
+					</li>
+					<li class=" tsd-kind-function tsd-parent-kind-module">
+						<a href="_favware_milky_tslint.html#isstring" class="tsd-kind-icon">is<wbr>String</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-module">
 						<a href="_favware_milky_tslint.html#milkylint" class="tsd-kind-icon">milky<wbr>Lint</a>
