+++ docs/modules/_favware_ytdl_prismplayer.html
@@ -159,21 +159,13 @@
 					<div class="tsd-signature tsd-kind-icon">Prism<wbr>Video<wbr>Format<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">YTDLVideoFormat</span><span class="tsd-signature-symbol"> &amp; </span><span class="tsd-signature-type">object</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/ytdl-prismplayer/src/index.ts#L10">packages/ytdl-prismplayer/src/index.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/ytdl-prismplayer/src/index.ts#L6">packages/ytdl-prismplayer/src/index.ts:6</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
-						<dl class="tsd-comment-tags">
-							<dt>typedef</dt>
-							<dd><p>PrismVideoFormat</p>
-							</dd>
-							<dt>summary</dt>
-							<dd><p>Custom type for ytdl videoFormat to include missing property</p>
-							</dd>
-							<dt>property</dt>
-							<dd><p>{number} [audio_sample_rate] - The sample rate of the audio format</p>
-							</dd>
-						</dl>
+						<div class="lead">
+							<p>Custom type for ytdl videoFormat to include missing property</p>
+						</div>
 					</div>
 				</section>
 			</section>
@@ -189,7 +181,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/favware/node-packages/blob/7b3cfca/packages/ytdl-prismplayer/src/index.ts#L46">packages/ytdl-prismplayer/src/index.ts:46</a></li>
+									<li>Defined in <a href="https://github.com/favware/node-packages/blob/2534ddd/packages/ytdl-prismplayer/src/index.ts#L38">packages/ytdl-prismplayer/src/index.ts:38</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -204,10 +196,15 @@
 								</li>
 								<li>
 									<h5><span class="tsd-flag ts-flagDefault value">Default value</span> options: <span class="tsd-signature-type">YTDLDownloadOptions</span><span class="tsd-signature-symbol"> =&nbsp;{}</span></h5>
+									<div class="tsd-comment tsd-typography">
+										<div class="lead">
+											<p>Any extra options to pass to ytdl, see <a href="https://github.com/fent/node-ytdl-core/#ytdlurl-options" class="external">https://github.com/fent/node-ytdl-core/#ytdlurl-options</a></p>
+										</div>
+									</div>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">Readable</span><span class="tsd-signature-symbol">&gt;</span></h4>
-							<p>Readable Stream (<a href="https://nodejs.org/api/stream.html#stream_class_stream_readable" class="external">https://nodejs.org/api/stream.html#stream_class_stream_readable</a>)</p>
+							<p>The playing song in a readable stream, see <a href="https://nodejs.org/api/stream.html#stream_class_stream_readable " class="external"> NodeJS.ReadableStream</a></p>
 						</li>
 					</ul>
 				</section>
@@ -226,10 +223,16 @@
 						<a href="_favware_crypto.html">@favware/crypto</a>
 					</li>
 					<li class=" tsd-kind-module">
+						<a href="_favware_eslint_config.html">@favware/eslint-<wbr>config</a>
+					</li>
+					<li class=" tsd-kind-module">
 						<a href="_favware_milky_tslint.html">@favware/milky-<wbr>tslint</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="_favware_querystring.html">@favware/querystring</a>
+					</li>
+					<li class=" tsd-kind-module">
+						<a href="_favware_tslint_config.html">@favware/tslint-<wbr>config</a>
 					</li>
 					<li class=" tsd-kind-module">
 						<a href="_favware_unescape.html">@favware/unescape</a>
