+++ packages/milky-tslint/src/events.ts
@@ -3,21 +3,13 @@ import { EventEmitter } from 'events';
 import { TSLintFile } from './types';
 import PluginError from 'plugin-error';
 
-/**
- * @typdef MilkyEvents Events output by this library
- */
+/** MilkyEvents Events output by this library */
 export interface MilkyEvents {
-  /**
-   * Fired where there is linted data
-   */
+  /** Fired where there is linted data */
   data: (file: TSLintFile) => void;
-  /**
-   * Fired when an error occurs during linting
-   */
+  /** Fired when an error occurs during linting */
   error: (err: PluginError) => void;
-  /**
-   * Fired when lintin is complete
-   */
+  /** Fired when lintin is complete */
   end: () => void;
 }
