+++ packages/speedtest/src/index.ts
@@ -4,4 +4,3 @@ export * from './speedtest';
 export * from './events';
 export * from './types';
 export default speedtest;
-module.exports = speedtest;
