+++ packages/unescape/src/unescape.ts
@@ -2,11 +2,7 @@ import {charSets} from '.';
 
 let regexCache: RegExp | null = null;
 
-/**
- * Transforms Character Set to regex for string replacement
- * @returns {RegExp}
- * @private
- */
+/** Transforms Character Set to regex for string replacement */
 const toRegex = (): RegExp => {
   if (regexCache) return regexCache;
 
@@ -18,46 +14,38 @@ const toRegex = (): RegExp => {
   return regex;
 };
 
-/**
- * Checks if the input is an Array<string> or not
- * @param {unknown} value The array to check if it only has strings
- * @returns {boolean}
- * @private
- */
-const isArrayOfOnlyStrings = (value: unknown): boolean => {
-  if (Array.isArray(value)) {
-    return Boolean(value.every(item => typeof item === 'string'));
+/** TypeGuard to ensure input is an array of only strings */
+const isArrayOfOnlyStrings = (array: unknown): array is string[] => {
+  if (Array.isArray(array)) {
+    return array.every(val => typeof val === 'string');
   }
 
   return false;
 };
 
+
 /**
  * Unescapes any given string
- * @method
- * @name unescape
- * @param {string | string[] | TemplateStringsArray} input input string to unescape
- * @param {any[]} values rest values
- * @returns {string} Unescape string
- * @example
  * ```ts
  * console.log(unescape('&lt;div&gt;abc&lt;/div&gt;'));
  * ```
- * @example
  * ```ts
  * // use template tags
  * console.log(unescape`&lt;div&gt;abc&lt;/div&gt;`);
  * ```
- * @example
  * ```ts
  * // or with variables
  * const unescaped = '&lt;div&gt;unescaped&lt;/div&gt;';
  * console.log(unescape`awesome text with some ${unescaped} section`);
  * ```
+ *
+ * @param input input string to unescape
+ * @param values rest values
+ * @returns Unescaped string
  */
 export const unescape = (input: string | string[] | TemplateStringsArray, ...values: string[]): string => {
   const regex: RegExp = toRegex();
-  const inputArray: string[] = isArrayOfOnlyStrings(input) ? input as string[] : [];
+  const inputArray: string[] = isArrayOfOnlyStrings(input) ? input : [];
   let str = '';
 
   if (inputArray.length) {
