+++ packages/yamlreader/src/index.ts
@@ -1,12 +1,8 @@
 import fs from 'fs';
 import yaml, { DumpOptions } from 'js-yaml';
 
-/**
- * FS Options
- * @typedef {FsOptions} fsOptions Options for the file reader
- * @property
- */
-interface FsOptions {
+/** FS Options */
+export interface FsOptions {
   encoding?:
   | 'utf8'
   | 'ascii'
@@ -21,28 +17,14 @@ interface FsOptions {
 
 /**
  * YAMLReaderOptions for configuring both `fs` and `js-yaml`
- * See [js-yaml configuration](https://github.com/nodeca/js-yaml#safeload-string---options-) for more details
- * Also see {@link fsOptions}
- * @typedef {YamlReaderOptions} YamlReaderOptions Options for the YAML parser
- * @property {number} [indent=2] indentation width to use (in spaces).
- * @property {boolean} [noArrayIndent=false] when true, will not add an indentation level to array elements
- * @property {boolean} [skipInvalid=false] do not throw on invalid types (like function in the safe schema) and skip pairs and single values with such types
- * @property {number} [flowLevel=-1] specifies level of nesting, when to switch from block to flow style for collections. -1 means block style everwhere
- * @property {string} [styles] Each tag may have own set of styles.  - "tag" => "style" map.
- * @property {SchemaDefinition} [schema=YAML.DEFAULT_SAFE_SCHEMA] specifies a schema to use
- * @property {boolean} [sortKeys=false] if `true`, sort keys when dumping YAML. If a function, use the function to sort the keys
- * @property {number} [lineWidth=80] set max line width
- * @property {boolean} [noRefs=false] if `true`, don't convert duplicate objects into references
- * @property {boolean} [noCompatMode=false] if `true` don't try to be compatible with older yaml versions. Currently: don't quote "yes", "no" and so on, as required for YAML 1.1
- * @property {boolean} [condenseFlow=false] if `true` flow sequences will be condensed, omitting the space between `a, b`. Eg. `'[a,b]'`
+ * See {@link https://github.com/nodeca/js-yaml#safeload-string---options- | js-yaml configuration} and {@link FsOptions} for details
  */
 export type YamlReaderOptions = DumpOptions & FsOptions;
 
 export class YamlReaderError extends Error {
   /**
-   * Create an YamlReaderError
+   * Create a YamlReaderError
    * @param {string} message The message the error should show
-   * @private
    */
   constructor(message: string) {
     super(message);
@@ -52,17 +34,14 @@ export class YamlReaderError extends Error {
 }
 
 /**
- * @method
- * @name readYaml
+ * Reads a YAML file and outputs a JavaScript object
  *
- * @param {string} filepath Path to the yaml file to parse
- * @param {YamlReaderOptions} [options] Options for the YAML parser and fs file reader
- * @returns {object} JSON-fied data
- * @example
  * ```ts
  * const data = readYaml('/path/to/some/yaml/file');
- * // => JavaScript object with the keys from the yaml
  * ```
+ *
+ * @param filepath Path to the yaml file to parse
+ * @param options Options for the YAML parser and fs file reader
  */
 export const readYaml = (filepath: string, options?: YamlReaderOptions) => {
   let str;
