+++ packages/zalgo/src/index.ts
@@ -2,4 +2,6 @@ import zalgo from './zalgo';
 
 export * from './zalgo';
 export * from './chars';
+export * from './interfaces';
+export * from './utils';
 export default zalgo;
