+++ packages/zalgo/src/interfaces.ts
@@ -1,0 +1,37 @@
+/** Interface for the characters object */
+export interface Chars {
+  /** All characters */
+  all?: string[];
+  /** Characters going up */
+  up: string[];
+  /** Characters in the middle */
+  middle: string[];
+  /** Characters going below */
+  down: string[];
+  /** Stored RegExp pattern */
+  pattern?: RegExp;
+
+  /** Index Signature for TS parsing */
+  [indexSignature: string]: string[] | RegExp | undefined;
+}
+
+/** Options for the Zalgo screw up */
+export interface ZalgoOptions {
+  /** Whether the zalgo should go up */
+  up?: boolean;
+  /** Whether the zalgo should go in the middle */
+  middle?: boolean;
+  /** Whether the zalgo should go downards */
+  down?: boolean;
+  /** Whether the zalgo should be mini or maxi */
+  size?: 'mini' | 'maxi' | '';
+}
+
+/** Counts map for the zalgo */
+export interface ZalgoCountsMap {
+  up: number;
+  middle: number;
+  down: number;
+
+  [indexSingature: string]: number;
+}
